estimatr

Fast procedures for small set of commonly-used, design-appropriate estimators with robust standard errors and confidence intervals. Includes estimators for linear regression, instrumental variables regression, difference-in-means, Horvitz-Thompson estimation, and regression improving precision of experimental estimates by interacting treatment with centered pre-treatment covariates introduced by Lin (2013) <doi:10.1214/12-AOAS583>.

See also

Author

Maintainer: Graeme Blair graeme.blair@gmail.com

Authors:

Jasper Cooper jjc2247@columbia.edu

Alexander Coppock alex.coppock@yale.edu

Macartan Humphreys macartan@gmail.com

Luke Sonnet luke.sonnet@gmail.com

Other contributors:

Neal Fultz nfultz@gmail.com [contributor]

Lily Medina lilymiru@gmail.com [contributor]

Russell Lenth russell-lenth@uiowa.edu [contributor]