Design-based difference-in-means estimator

This difference-in-means estimator selects the appropriate point estimator, standard errors, and degrees of freedom for a variety of designs: unit randomized, cluster randomized, block randomized, block-cluster randomized, matched-pairs, and matched-pair cluster randomized designs

difference_in_means(formula, data, blocks, clusters, weights, subset,
  condition1 = NULL, condition2 = NULL, ci = TRUE, alpha = 0.05)

Arguments

formula
an object of class formula, as in lm, such as Y ~ Z with only one variable on the right-hand side, the treatment.

data
A data.frame.

blocks
An optional bare (unquoted) name of the block variable. Use for blocked designs only.

clusters
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data; used for cluster randomized designs. For blocked designs, clusters must nest within blocks.

weights
the bare (unquoted) names of the weights variable in the supplied data.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

condition1
value in the treatment vector of the condition to be the control. Effects are estimated with condition1 as the control and condition2 as the treatment. If unspecified, condition1 is the "first" condition and condition2 is the "second" according to levels if the treatment is a factor or according to a sortif it is a numeric or character variable (i.e if unspecified and the treatment is 0s and 1s, condition1 will by default be 0 and condition2 will be 1). See the examples for more.

condition2
value in the treatment vector of the condition to be the treatment. See condition1.

ci
A boolean for whether to compute and return p-values and confidence intervals, TRUE by default.

alpha
The significance level, 0.05 by default.

Value

difference_in_means returns an object of class "difference_in_means".

The functions summary and tidy can be used to get the results as a data.frame. To get useful data out of the return, you can use these data frames, you can use the resulting list directly, or you can use the generic accessor functions coef, andconfint.

An object of class "difference_in_means" is a list containing at least the following components:

est

the estimated coefficients

se

the estimated standard errors

df

the estimated degrees of freedom

p

the p-values from the t-test using est, se, and df

ci_lower

the lower bound of the 1 - alpha percent confidence interval

ci_upper

the upper bound of the 1 - alpha percent confidence interval

coefficient_name

a character vector of coefficient names

alpha

the significance level specified by the user

N

the number of observations used

outcome

the name of the outcome variable

design

the name of the design learned from the arguments passed

Details

This function implements a difference-in-means estimator, with support for blocked, clustered, matched-pairs, block-clustered, and matched-pair clustered designs. One specifies their design by passing the blocks and clusters in their data and this function chooses which estimator is most appropriate.

If you pass only blocks, if all blocks are of size two, we will infer that the design is a matched-pairs design. If they are all size four or larger, we will infer that it is a regular blocked design. If you pass both blocks and clusters, we will similarly infer whether it is a matched-pairs clustered design or a block-clustered design the number of clusters per block. If the user passes only clusters, we will infer that the design was cluster-randomized. If the user specifies neither the blocks nor the clusters, a regular Welch's t-test will be performed.

Importantly, if the user specifies weights, the estimation is handed off to lm_robust with the appropriate robust standard errors as weighted difference-in-means estimators are not implemented here. More details of the about each of the estimators can be found in the technical notes.

References

Gerber, Alan S, and Donald P Green. 2012. Field Experiments: Design, Analysis, and Interpretation. New York: W.W. Norton. Imai, Kosuke, Gary King, Clayton Nall. 2009. "The Essential Role of Pair Matching in Cluster-Randomized Experiments, with Application to the Mexican Universal Health Insurance Evaluation." Statistical Science 24 (1). Institute of Mathematical Statistics: 29-53. https://doi.org/10.1214/08-STS274.

Examples

library(fabricatr)
library(randomizr)
# Get appropriate standard errors for simple designs
dat <- fabricate(
  N = 100,
  Y = rnorm(100),
  Z_simp = simple_ra(N, prob = 0.4),
)

table(dat$Z_simp)

#> 
#>  0  1 
#> 51 49 

difference_in_means(Y ~ Z_simp, data = dat)

#> [1] "Design: Standard"
#>   coefficient_name       est        se        p   ci_lower  ci_upper       df
#> 1           Z_simp -0.110078 0.2005905 0.584412 -0.5081456 0.2879895 97.95751
#>   outcome
#> 1       Y

# Accurates estimates and standard errors for clustered designs
dat$clust <- sample(20, size = nrow(dat), replace = TRUE)
dat$Z_clust <- cluster_ra(dat$clust, prob = 0.6)

table(dat$Z_clust, dat$clust)

#>    
#>     1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20
#>   0 3 5 7 0 0 0 0 8 5  0  0  0  0  0  2  0  0  6  0  6
#>   1 0 0 0 7 7 6 2 0 0  4  6  1  8  3  0  3  6  0  5  0

summary(difference_in_means(Y ~ Z_clust, clusters = clust, data = dat))

#> $coefficients
#>           Estimate Std. Error  Pr(>|t|)   ci_lower  ci_upper       df
#> Z_clust 0.09426617  0.2147297 0.6676325 -0.3680591 0.5565915 13.44919
#> 
#> $design
#> [1] "Clustered"
#> 

# Accurate estimates and standard errors for blocked designs
dat$block <- rep(1:10, each = 10)
dat$Z_block <- block_ra(dat$block, prob = 0.5)

table(dat$Z_block, dat$block)

#>    
#>     1 2 3 4 5 6 7 8 9 10
#>   0 5 5 5 5 5 5 5 5 5  5
#>   1 5 5 5 5 5 5 5 5 5  5

difference_in_means(Y ~ Z_block, blocks = block, data = dat)

#> [1] "Design: Blocked"
#>   coefficient_name       est        se         p   ci_lower  ci_upper df
#> 1          Z_block -0.075066 0.2077237 0.7187723 -0.4884493 0.3383173 80
#>   outcome
#> 1       Y

# Matched-pair estimates and standard errors are also accurate
# Specified same as blocked design, function learns that
# it is matched pair from size of blocks!
dat$pairs <- rep(1:50, each = 2)
dat$Z_pairs <- block_ra(dat$pairs, prob = 0.5)

table(dat$pairs, dat$Z_pairs)

#>     
#>      0 1
#>   1  1 1
#>   2  1 1
#>   3  1 1
#>   4  1 1
#>   5  1 1
#>   6  1 1
#>   7  1 1
#>   8  1 1
#>   9  1 1
#>   10 1 1
#>   11 1 1
#>   12 1 1
#>   13 1 1
#>   14 1 1
#>   15 1 1
#>   16 1 1
#>   17 1 1
#>   18 1 1
#>   19 1 1
#>   20 1 1
#>   21 1 1
#>   22 1 1
#>   23 1 1
#>   24 1 1
#>   25 1 1
#>   26 1 1
#>   27 1 1
#>   28 1 1
#>   29 1 1
#>   30 1 1
#>   31 1 1
#>   32 1 1
#>   33 1 1
#>   34 1 1
#>   35 1 1
#>   36 1 1
#>   37 1 1
#>   38 1 1
#>   39 1 1
#>   40 1 1
#>   41 1 1
#>   42 1 1
#>   43 1 1
#>   44 1 1
#>   45 1 1
#>   46 1 1
#>   47 1 1
#>   48 1 1
#>   49 1 1
#>   50 1 1

difference_in_means(Y ~ Z_pairs, blocks = pairs, data = dat)

#> [1] "Design: Matched-pair"
#>   coefficient_name       est        se         p   ci_lower  ci_upper df
#> 1          Z_pairs 0.1279107 0.2035161 0.5325926 -0.2810702 0.5368916 49
#>   outcome
#> 1       Y

# Also works with multi-valued treatments if users specify
# comparison of interest
dat$Z_multi <- simple_ra(
  nrow(dat),
  condition_names = c("Treatment 2", "Treatment 1", "Control"),
  prob_each = c(0.4, 0.4, 0.2)
)

# Only need to specify which condition is treated "condition2" and
# which is control "condition1"
difference_in_means(
  Y ~ Z_multi,
  condition1 = "Treatment 2",
  condition2 = "Control",
  data = dat
)

#> [1] "Design: Standard"
#>   coefficient_name         est        se        p   ci_lower  ci_upper       df
#> 1   Z_multiControl -0.02186428 0.3002911 0.942377 -0.6317747 0.5880462 34.54505
#>   outcome
#> 1       Y

difference_in_means(
  Y ~ Z_multi,
  condition1 = "Treatment 1",
  condition2 = "Control",
  data = dat
)

#> [1] "Design: Standard"
#>   coefficient_name       est        se         p   ci_lower  ci_upper       df
#> 1   Z_multiControl 0.2074861 0.2972427 0.4900263 -0.3971726 0.8121447 33.12422
#>   outcome
#> 1       Y

# Specifying weights will result in estimation via `lm_robust`
dat$w <- runif(nrow(dat))
difference_in_means(Y ~ Z_simp, weights = w, data = dat)

#> [1] "Design: Standard (weighted)"
#>   coefficient_name         est       se         p   ci_lower  ci_upper df
#> 1           Z_simp -0.08534263 0.225205 0.7055409 -0.5322546 0.3615694 98
#>   outcome
#> 1       Y

lm_robust(Y ~ Z_simp, weights = w, data = dat)

#>   coefficient_name         est        se         p   ci_lower  ci_upper df
#> 1      (Intercept)  0.02889963 0.1598783 0.8569291 -0.2883737 0.3461729 98
#> 2           Z_simp -0.08534263 0.2252050 0.7055409 -0.5322546 0.3615694 98
#>   outcome
#> 1       Y
#> 2       Y