Built-in Estimators: Difference-in-means

Built-in Estimators: Difference-in-means

difference_in_means(formula, block_variable_name = NULL,
  cluster_variable_name = NULL, condition1 = NULL, condition2 = NULL,
  data, weights = NULL, subset = NULL, alpha = 0.05)

Arguments

formula
An object of class "formula", such as Y ~ Z

block_variable_name
An optional bare (unquoted) name of the block variable. Use for blocked designs only.

cluster_variable_name
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data; used for cluster randomized designs. For blocked designs, clusters must be within blocks.

condition1
names of the conditions to be compared. Effects are estimated with condition1 as control and condition2 as treatment. If unspecified, condition1 is the "first" condition and condition2 is the "second" according to r defaults.

condition2
names of the conditions to be compared. Effects are estimated with condition1 as control and condition2 as treatment. If unspecified, condition1 is the "first" condition and condition2 is the "second" according to r defaults.

data
A data.frame.

weights
An optional bare (unquoted) name of the weights variable.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

alpha
The significance level, 0.05 by default.

Details

This function implements difference-in-means estimation, with and without blocking. Standard errors are estimated as the square root of the sum of the within-group variances, divided by their respective sample sizes (Equation 3.6 in Gerber and Green 2012). If blocked, the difference in means estimate is taken in each block, then averaged together according to block size.

Examples

 df <- data.frame(Y = rnorm(100),
                  Z = sample(1:3, 100, replace = TRUE),
                  block = sample(c("A", "B", "C"), 100, replace = TRUE))

 difference_in_means(Y ~ Z, data = df)

#>   coefficient_name          est           se            p      ci_lower
#> 1                Z 0.1224874721 0.2674281763 0.6485811709 -0.4123727644
#>       ci_upper          df
#> 1 0.6573477086 60.41607609

 difference_in_means(Y ~ Z, condition1 = 3, condition2 = 2, data = df)

#>   coefficient_name          est           se            p      ci_lower
#> 1                Z 0.2770965566 0.2432305921 0.2588673272 -0.2088470169
#>       ci_upper          df
#> 1 0.7630401301 63.76421495

 difference_in_means(Y ~ Z, block_variable_name = block, data = df)

#>   coefficient_name           est           se            p     ci_lower
#> 1                Z 0.02735372352 0.2960366265 0.9265709693 -0.560197033
#>       ci_upper df
#> 1 0.6149044801 97

 difference_in_means(Y ~ Z, block_variable_name = block, condition1 = 3, condition2 = 2, data = df)

#>   coefficient_name          est           se            p      ci_lower
#> 1                Z 0.2107775019 0.2830092347 0.4582099063 -0.3509174881
#>       ci_upper df
#> 1 0.7724724918 97