Generate condition probability matrix given clusters and probabilities

Generate condition probability matrix given clusters and probabilities

Usage

gen_pr_matrix_cluster(clusters, treat_probs, simple)

Arguments

clusters

A vector of clusters

treat_probs

A vector of treatment (condition 2) probabilities

simple

A boolean for whether the assignment is a random sample assignment (TRUE, default) or complete random assignment (FALSE)

Value

a numeric 2n*2n matrix of marginal and joint condition treatment probabilities to be passed to the condition_pr_mat argument of horvitz_thompson.

See also