Horvitz-Thompson estimator of treatment effects

Horvitz-Thompson estimator of treatment effects

horvitz_thompson(formula, condition_pr_variable_name = NULL,
  condition_pr_matrix = NULL, block_variable_name = NULL,
  cluster_variable_name = NULL, declaration = NULL, estimator = "ht",
  constant_effects = FALSE, data, subset = NULL, alpha = 0.05,
  condition1 = NULL, condition2 = NULL)

Arguments

formula
An object of class "formula", such as Y ~ Z

condition_pr_variable_name
A bare (unquoted) name of the variable with the condition 2 (treatment) probabilities.

condition_pr_matrix
A 2n * 2n matrix of marginal and joint probabilities of all units in condition1 and condition2; see details.

block_variable_name
An optional bare (unquoted) name of the block variable. Use for blocked designs only.

cluster_variable_name
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data; used for cluster randomized designs. For blocked designs, clusters must be within blocks.

declaration
An object of class "ra_declaration", from the randomizr package. Can substitute for manually specifying clusters, blocks, and condition probabilities.

constant_effects
should constant effects be assumed to estimate the covariance between potential outcomes, default = FALSE.

data
A data.frame.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

alpha
The significance level, 0.05 by default.

condition1
names of the conditions to be compared. Effects are estimated with condition1 as control and condition2 as treatment. If unspecified, condition1 is the "first" condition and condition2 is the "second" according to r defaults.

condition2
names of the conditions to be compared. Effects are estimated with condition1 as control and condition2 as treatment. If unspecified, condition1 is the "first" condition and condition2 is the "second" according to r defaults.

Details

This function implements the Horvitz-Thompson estimator for treatment effects.