Horvitz-Thompson estimator of treatment effects

Horvitz-Thompson estimator of treatment effects

horvitz_thompson(formula, data, condition_prs, blocks, clusters,
  condition_pr_mat = NULL, declaration = NULL, subset,
  se_type = c("youngs", "constant"), collapsed = FALSE, alpha = 0.05,
  condition1 = NULL, condition2 = NULL)

Arguments

formula
An object of class "formula", such as Y ~ Z

data
A data.frame.

condition_prs
An optional bare (unquoted) name of the variable with the condition 2 (treatment) probabilities.

blocks
An optional bare (unquoted) name of the block variable. Use for blocked designs only.

clusters
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data; used for cluster randomized designs. For blocked designs, clusters must be within blocks.

condition_pr_mat
An optional 2n * 2n matrix of marginal and joint probabilities of all units in condition1 and condition2, can be used in place of condition_prs. See details.

declaration
An object of class "ra_declaration", from the randomizr package that is an alternative way of specifying the design. Cannot be used along with any of condition_prs, blocks, clusters, or condition_pr_mat. See details.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

se_type
can be one of c("youngs", "constant") and correspond's to estimating the standard errors using Young's inequality (default, conservative), or the constant effects assumption.

collapsed
A boolean used to collapse clusters to their cluster totals for variance estimation, FALSE by default.

alpha
The significance level, 0.05 by default.

condition1
names of the conditions to be compared. Effects are estimated with condition1 as control and condition2 as treatment. If unspecified, condition1 is the "first" condition and condition2 is the "second" according to r defaults.

condition2
names of the conditions to be compared. Effects are estimated with condition1 as control and condition2 as treatment. If unspecified, condition1 is the "first" condition and condition2 is the "second" according to r defaults.

Details

This function implements the Horvitz-Thompson estimator for treatment effects.

Examples

# Simulate data
n <- 10
dat <- data.frame(y = rnorm(n))

#----------
# Simple random assignment
#----------
dat$p <- 0.5
dat$z <- rbinom(n, size = 1, prob = dat$p)

# If you only pass condition_prs, we assume simple random sampling
horvitz_thompson(y ~ z, data = dat, condition_prs = p)

#>              y t condition_probabilities
#> 1  -0.42938009 0                     0.5
#> 2   1.36046133 0                     0.5
#> 3  -0.07085743 0                     0.5
#> 4  -0.27215368 0                     0.5
#> 5  -2.44668003 1                     0.5
#> 6   0.06548664 0                     0.5
#> 7  -1.09850890 0                     0.5
#> 8  -0.63317818 1                     0.5
#> 9  -2.06365445 1                     0.5
#> 10  2.64893203 1                     0.5

#>   coefficient_name        est       se         p  ci_lower ci_upper df
#> 1                z -0.4099257 0.916253 0.6664566 -2.522809 1.702958  8

# Assume constant effects instead
horvitz_thompson(y ~ z, data = dat, condition_prs = p, se_type = "constant")

#>              y t condition_probabilities
#> 1  -0.42938009 0                     0.5
#> 2   1.36046133 0                     0.5
#> 3  -0.07085743 0                     0.5
#> 4  -0.27215368 0                     0.5
#> 5  -2.44668003 1                     0.5
#> 6   0.06548664 0                     0.5
#> 7  -1.09850890 0                     0.5
#> 8  -0.63317818 1                     0.5
#> 9  -2.06365445 1                     0.5
#> 10  2.64893203 1                     0.5

#>   coefficient_name        est        se         p  ci_lower ci_upper df
#> 1                z -0.4099257 0.9070368 0.6633143 -2.501556 1.681705  8

# Also can use randomizr to pass a declaration
srs_declaration <- randomizr::declare_ra(N = nrow(dat), prob = 0.5, simple = TRUE)
horvitz_thompson(y ~ z, data = dat, declaration = srs_declaration)

#>              y t condition_probabilities
#> 1  -0.42938009 0                     0.5
#> 2   1.36046133 0                     0.5
#> 3  -0.07085743 0                     0.5
#> 4  -0.27215368 0                     0.5
#> 5  -2.44668003 1                     0.5
#> 6   0.06548664 0                     0.5
#> 7  -1.09850890 0                     0.5
#> 8  -0.63317818 1                     0.5
#> 9  -2.06365445 1                     0.5
#> 10  2.64893203 1                     0.5

#>   coefficient_name        est       se         p  ci_lower ci_upper df
#> 1                z -0.4099257 0.916253 0.6664566 -2.522809 1.702958  8

#----------
# Complete random assignemtn
#----------

dat$z <- sample(rep(0:1, each = n/2))
# Can use a declaration
crs_declaration <- randomizr::declare_ra(N = nrow(dat), prob = 0.5, simple = FALSE)
horvitz_thompson(y ~ z, data = dat, declaration = crs_declaration)

#>              y t condition_probabilities
#> 1  -0.42938009 1                     0.5
#> 2   1.36046133 0                     0.5
#> 3  -0.07085743 1                     0.5
#> 4  -0.27215368 1                     0.5
#> 5  -2.44668003 0                     0.5
#> 6   0.06548664 1                     0.5
#> 7  -1.09850890 0                     0.5
#> 8  -0.63317818 1                     0.5
#> 9  -2.06365445 0                     0.5
#> 10  2.64893203 0                     0.5
#> [1] "full youngs"
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#>  [2,] 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#>  [3,] 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421
#>  [4,] 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421
#>  [5,] 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421
#>  [6,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421
#>  [7,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000
#>  [8,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#>  [9,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [10,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [11,] 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#> [12,] 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#> [13,] 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579
#> [14,] 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579
#> [15,] 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579
#> [16,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579
#> [17,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000
#> [18,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#> [19,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#> [20,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,] 0.2368421 0.2368421 0.2368421 0.0000000 0.2631579 0.2631579 0.2631579
#>  [2,] 0.2368421 0.2368421 0.2368421 0.2631579 0.0000000 0.2631579 0.2631579
#>  [3,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.0000000 0.2631579
#>  [4,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.0000000
#>  [5,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#>  [6,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#>  [7,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#>  [8,] 0.5000000 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#>  [9,] 0.2368421 0.5000000 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#> [10,] 0.2368421 0.2368421 0.5000000 0.2631579 0.2631579 0.2631579 0.2631579
#> [11,] 0.2631579 0.2631579 0.2631579 0.5000000 0.2368421 0.2368421 0.2368421
#> [12,] 0.2631579 0.2631579 0.2631579 0.2368421 0.5000000 0.2368421 0.2368421
#> [13,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.5000000 0.2368421
#> [14,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.5000000
#> [15,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [16,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [17,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [18,] 0.0000000 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [19,] 0.2631579 0.0000000 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [20,] 0.2631579 0.2631579 0.0000000 0.2368421 0.2368421 0.2368421 0.2368421
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]
#>  [1,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [2,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [3,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [4,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [5,] 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [6,] 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579
#>  [7,] 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579
#>  [8,] 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579
#>  [9,] 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579
#> [10,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000
#> [11,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [12,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [13,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [14,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [15,] 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [16,] 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421
#> [17,] 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421
#> [18,] 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421
#> [19,] 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421
#> [20,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000

#>   coefficient_name        est       se         p  ci_lower ci_upper df
#> 1                z 0.05187346 0.931617 0.9569614 -2.096439 2.200186  8

# Can precompute condition_pr_mat and pass it
# (faster for multiple runs with same condition probability matrix)
crs_pr_mat <- declaration_to_condition_pr_mat(crs_declaration)
horvitz_thompson(y ~ z, data = dat, condition_pr_mat = crs_pr_mat)

#>              y t condition_probabilities
#> 1  -0.42938009 1                     0.5
#> 2   1.36046133 0                     0.5
#> 3  -0.07085743 1                     0.5
#> 4  -0.27215368 1                     0.5
#> 5  -2.44668003 0                     0.5
#> 6   0.06548664 1                     0.5
#> 7  -1.09850890 0                     0.5
#> 8  -0.63317818 1                     0.5
#> 9  -2.06365445 0                     0.5
#> 10  2.64893203 0                     0.5
#> [1] "full youngs"
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#>  [2,] 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#>  [3,] 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421
#>  [4,] 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421
#>  [5,] 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421
#>  [6,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421
#>  [7,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000
#>  [8,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#>  [9,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [10,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [11,] 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#> [12,] 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#> [13,] 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579
#> [14,] 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579
#> [15,] 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579
#> [16,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579
#> [17,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000
#> [18,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#> [19,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#> [20,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,] 0.2368421 0.2368421 0.2368421 0.0000000 0.2631579 0.2631579 0.2631579
#>  [2,] 0.2368421 0.2368421 0.2368421 0.2631579 0.0000000 0.2631579 0.2631579
#>  [3,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.0000000 0.2631579
#>  [4,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.0000000
#>  [5,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#>  [6,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#>  [7,] 0.2368421 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#>  [8,] 0.5000000 0.2368421 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#>  [9,] 0.2368421 0.5000000 0.2368421 0.2631579 0.2631579 0.2631579 0.2631579
#> [10,] 0.2368421 0.2368421 0.5000000 0.2631579 0.2631579 0.2631579 0.2631579
#> [11,] 0.2631579 0.2631579 0.2631579 0.5000000 0.2368421 0.2368421 0.2368421
#> [12,] 0.2631579 0.2631579 0.2631579 0.2368421 0.5000000 0.2368421 0.2368421
#> [13,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.5000000 0.2368421
#> [14,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.5000000
#> [15,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [16,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [17,] 0.2631579 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [18,] 0.0000000 0.2631579 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [19,] 0.2631579 0.0000000 0.2631579 0.2368421 0.2368421 0.2368421 0.2368421
#> [20,] 0.2631579 0.2631579 0.0000000 0.2368421 0.2368421 0.2368421 0.2368421
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]
#>  [1,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [2,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [3,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [4,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [5,] 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579
#>  [6,] 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579 0.2631579
#>  [7,] 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579 0.2631579
#>  [8,] 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579 0.2631579
#>  [9,] 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000 0.2631579
#> [10,] 0.2631579 0.2631579 0.2631579 0.2631579 0.2631579 0.0000000
#> [11,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [12,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [13,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [14,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [15,] 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421
#> [16,] 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421 0.2368421
#> [17,] 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421 0.2368421
#> [18,] 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421 0.2368421
#> [19,] 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000 0.2368421
#> [20,] 0.2368421 0.2368421 0.2368421 0.2368421 0.2368421 0.5000000

#>   coefficient_name        est       se         p  ci_lower ci_upper df
#> 1                z 0.05187346 0.931617 0.9569614 -2.096439 2.200186  8

#----------
# More complicated assignment
#----------

# arbitrary permutation matrix
possible_treats <- cbind(
  c(1, 1, 0, 1, 0, 0, 0, 1, 1, 0),
  c(0, 1, 1, 0, 1, 1, 0, 1, 0, 1),
  c(1, 0, 1, 1, 1, 1, 1, 0, 0, 0)
)
arb_pr_mat <- permutations_to_condition_pr_mat(possible_treats)
# Simulating a column to be realized treatment
dat$z <- possible_treats[, sample(ncol(possible_treats), size = 1)]
horvitz_thompson(y ~ z, data = dat, condition_pr_mat = arb_pr_mat)

#>              y t condition_probabilities
#> 1  -0.42938009 0               0.3333333
#> 2   1.36046133 1               0.3333333
#> 3  -0.07085743 1               0.3333333
#> 4  -0.27215368 0               0.3333333
#> 5  -2.44668003 1               0.3333333
#> 6   0.06548664 1               0.3333333
#> 7  -1.09850890 0               0.6666667
#> 8  -0.63317818 1               0.3333333
#> 9  -2.06365445 0               0.6666667
#> 10  2.64893203 1               0.6666667
#> [1] "full youngs"
#>            0_1       0_2       0_3       0_4       0_5       0_6       0_7
#> 0_1  0.6666667 0.3333333 0.3333333 0.6666667 0.3333333 0.3333333 0.3333333
#> 0_2  0.3333333 0.6666667 0.3333333 0.3333333 0.3333333 0.3333333 0.0000000
#> 0_3  0.3333333 0.3333333 0.6666667 0.3333333 0.6666667 0.6666667 0.3333333
#> 0_4  0.6666667 0.3333333 0.3333333 0.6666667 0.3333333 0.3333333 0.3333333
#> 0_5  0.3333333 0.3333333 0.6666667 0.3333333 0.6666667 0.6666667 0.3333333
#> 0_6  0.3333333 0.3333333 0.6666667 0.3333333 0.6666667 0.6666667 0.3333333
#> 0_7  0.3333333 0.0000000 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#> 0_8  0.3333333 0.6666667 0.3333333 0.3333333 0.3333333 0.3333333 0.0000000
#> 0_9  0.3333333 0.3333333 0.0000000 0.3333333 0.0000000 0.0000000 0.0000000
#> 0_10 0.0000000 0.3333333 0.3333333 0.0000000 0.3333333 0.3333333 0.0000000
#> 1_1  0.0000000 0.3333333 0.3333333 0.0000000 0.3333333 0.3333333 0.0000000
#> 1_2  0.3333333 0.0000000 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#> 1_3  0.3333333 0.3333333 0.0000000 0.3333333 0.0000000 0.0000000 0.0000000
#> 1_4  0.0000000 0.3333333 0.3333333 0.0000000 0.3333333 0.3333333 0.0000000
#> 1_5  0.3333333 0.3333333 0.0000000 0.3333333 0.0000000 0.0000000 0.0000000
#> 1_6  0.3333333 0.3333333 0.0000000 0.3333333 0.0000000 0.0000000 0.0000000
#> 1_7  0.3333333 0.6666667 0.3333333 0.3333333 0.3333333 0.3333333 0.0000000
#> 1_8  0.3333333 0.0000000 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#> 1_9  0.3333333 0.3333333 0.6666667 0.3333333 0.6666667 0.6666667 0.3333333
#> 1_10 0.6666667 0.3333333 0.3333333 0.6666667 0.3333333 0.3333333 0.3333333
#>            0_8       0_9      0_10       1_1       1_2       1_3       1_4
#> 0_1  0.3333333 0.3333333 0.0000000 0.0000000 0.3333333 0.3333333 0.0000000
#> 0_2  0.6666667 0.3333333 0.3333333 0.3333333 0.0000000 0.3333333 0.3333333
#> 0_3  0.3333333 0.0000000 0.3333333 0.3333333 0.3333333 0.0000000 0.3333333
#> 0_4  0.3333333 0.3333333 0.0000000 0.0000000 0.3333333 0.3333333 0.0000000
#> 0_5  0.3333333 0.0000000 0.3333333 0.3333333 0.3333333 0.0000000 0.3333333
#> 0_6  0.3333333 0.0000000 0.3333333 0.3333333 0.3333333 0.0000000 0.3333333
#> 0_7  0.0000000 0.0000000 0.0000000 0.0000000 0.3333333 0.0000000 0.0000000
#> 0_8  0.6666667 0.3333333 0.3333333 0.3333333 0.0000000 0.3333333 0.3333333
#> 0_9  0.3333333 0.3333333 0.0000000 0.0000000 0.0000000 0.3333333 0.0000000
#> 0_10 0.3333333 0.0000000 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333
#> 1_1  0.3333333 0.0000000 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333
#> 1_2  0.0000000 0.0000000 0.0000000 0.0000000 0.3333333 0.0000000 0.0000000
#> 1_3  0.3333333 0.3333333 0.0000000 0.0000000 0.0000000 0.3333333 0.0000000
#> 1_4  0.3333333 0.0000000 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333
#> 1_5  0.3333333 0.3333333 0.0000000 0.0000000 0.0000000 0.3333333 0.0000000
#> 1_6  0.3333333 0.3333333 0.0000000 0.0000000 0.0000000 0.3333333 0.0000000
#> 1_7  0.6666667 0.3333333 0.3333333 0.3333333 0.0000000 0.3333333 0.3333333
#> 1_8  0.0000000 0.0000000 0.0000000 0.0000000 0.3333333 0.0000000 0.0000000
#> 1_9  0.3333333 0.0000000 0.3333333 0.3333333 0.3333333 0.0000000 0.3333333
#> 1_10 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333 0.3333333 0.0000000
#>            1_5       1_6       1_7       1_8       1_9      1_10
#> 0_1  0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.6666667
#> 0_2  0.3333333 0.3333333 0.6666667 0.0000000 0.3333333 0.3333333
#> 0_3  0.0000000 0.0000000 0.3333333 0.3333333 0.6666667 0.3333333
#> 0_4  0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.6666667
#> 0_5  0.0000000 0.0000000 0.3333333 0.3333333 0.6666667 0.3333333
#> 0_6  0.0000000 0.0000000 0.3333333 0.3333333 0.6666667 0.3333333
#> 0_7  0.0000000 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333
#> 0_8  0.3333333 0.3333333 0.6666667 0.0000000 0.3333333 0.3333333
#> 0_9  0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333
#> 0_10 0.0000000 0.0000000 0.3333333 0.0000000 0.3333333 0.0000000
#> 1_1  0.0000000 0.0000000 0.3333333 0.0000000 0.3333333 0.0000000
#> 1_2  0.0000000 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333
#> 1_3  0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333
#> 1_4  0.0000000 0.0000000 0.3333333 0.0000000 0.3333333 0.0000000
#> 1_5  0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333
#> 1_6  0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333
#> 1_7  0.3333333 0.3333333 0.6666667 0.0000000 0.3333333 0.3333333
#> 1_8  0.0000000 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333
#> 1_9  0.0000000 0.0000000 0.3333333 0.3333333 0.6666667 0.3333333
#> 1_10 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.6666667

#>   coefficient_name       est       se         p  ci_lower ci_upper df
#> 1                z 0.9337886 1.303373 0.4941001 -2.071794 3.939371  8

# Clustered treatment, complete random assigment
# Simulating data
dat$cl <- rep(1:4, times = c(2, 2, 3, 3))
clust_crs_decl <- randomizr::declare_ra(N = nrow(dat), clust_var = dat$cl, prob = 0.5)
dat$z <- randomizr::conduct_ra(clust_crs_decl)
# Regular SE using Young's inequality
horvitz_thompson(y ~ z, data = dat, declaration = clust_crs_decl)

#>              y t clusters condition_probabilities
#> 1  -0.42938009 0        1                     0.5
#> 2   1.36046133 0        1                     0.5
#> 3  -0.07085743 0        2                     0.5
#> 4  -0.27215368 0        2                     0.5
#> 5  -2.44668003 1        3                     0.5
#> 6   0.06548664 1        3                     0.5
#> 7  -1.09850890 1        3                     0.5
#> 8  -0.63317818 1        4                     0.5
#> 9  -2.06365445 1        4                     0.5
#> 10  2.64893203 1        4                     0.5
#> [1] "full youngs"
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.5000000 0.5000000 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#>  [2,] 0.5000000 0.5000000 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#>  [3,] 0.1666667 0.1666667 0.5000000 0.5000000 0.1666667 0.1666667 0.1666667
#>  [4,] 0.1666667 0.1666667 0.5000000 0.5000000 0.1666667 0.1666667 0.1666667
#>  [5,] 0.1666667 0.1666667 0.1666667 0.1666667 0.5000000 0.5000000 0.5000000
#>  [6,] 0.1666667 0.1666667 0.1666667 0.1666667 0.5000000 0.5000000 0.5000000
#>  [7,] 0.1666667 0.1666667 0.1666667 0.1666667 0.5000000 0.5000000 0.5000000
#>  [8,] 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#>  [9,] 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#> [10,] 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#> [11,] 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#> [12,] 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#> [13,] 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333
#> [14,] 0.3333333 0.3333333 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333
#> [15,] 0.3333333 0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.0000000
#> [16,] 0.3333333 0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.0000000
#> [17,] 0.3333333 0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.0000000
#> [18,] 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#> [19,] 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#> [20,] 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,] 0.1666667 0.1666667 0.1666667 0.0000000 0.0000000 0.3333333 0.3333333
#>  [2,] 0.1666667 0.1666667 0.1666667 0.0000000 0.0000000 0.3333333 0.3333333
#>  [3,] 0.1666667 0.1666667 0.1666667 0.3333333 0.3333333 0.0000000 0.0000000
#>  [4,] 0.1666667 0.1666667 0.1666667 0.3333333 0.3333333 0.0000000 0.0000000
#>  [5,] 0.1666667 0.1666667 0.1666667 0.3333333 0.3333333 0.3333333 0.3333333
#>  [6,] 0.1666667 0.1666667 0.1666667 0.3333333 0.3333333 0.3333333 0.3333333
#>  [7,] 0.1666667 0.1666667 0.1666667 0.3333333 0.3333333 0.3333333 0.3333333
#>  [8,] 0.5000000 0.5000000 0.5000000 0.3333333 0.3333333 0.3333333 0.3333333
#>  [9,] 0.5000000 0.5000000 0.5000000 0.3333333 0.3333333 0.3333333 0.3333333
#> [10,] 0.5000000 0.5000000 0.5000000 0.3333333 0.3333333 0.3333333 0.3333333
#> [11,] 0.3333333 0.3333333 0.3333333 0.5000000 0.5000000 0.1666667 0.1666667
#> [12,] 0.3333333 0.3333333 0.3333333 0.5000000 0.5000000 0.1666667 0.1666667
#> [13,] 0.3333333 0.3333333 0.3333333 0.1666667 0.1666667 0.5000000 0.5000000
#> [14,] 0.3333333 0.3333333 0.3333333 0.1666667 0.1666667 0.5000000 0.5000000
#> [15,] 0.3333333 0.3333333 0.3333333 0.1666667 0.1666667 0.1666667 0.1666667
#> [16,] 0.3333333 0.3333333 0.3333333 0.1666667 0.1666667 0.1666667 0.1666667
#> [17,] 0.3333333 0.3333333 0.3333333 0.1666667 0.1666667 0.1666667 0.1666667
#> [18,] 0.0000000 0.0000000 0.0000000 0.1666667 0.1666667 0.1666667 0.1666667
#> [19,] 0.0000000 0.0000000 0.0000000 0.1666667 0.1666667 0.1666667 0.1666667
#> [20,] 0.0000000 0.0000000 0.0000000 0.1666667 0.1666667 0.1666667 0.1666667
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]
#>  [1,] 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#>  [2,] 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#>  [3,] 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#>  [4,] 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333
#>  [5,] 0.0000000 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333
#>  [6,] 0.0000000 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333
#>  [7,] 0.0000000 0.0000000 0.0000000 0.3333333 0.3333333 0.3333333
#>  [8,] 0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.0000000
#>  [9,] 0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.0000000
#> [10,] 0.3333333 0.3333333 0.3333333 0.0000000 0.0000000 0.0000000
#> [11,] 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#> [12,] 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#> [13,] 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#> [14,] 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667 0.1666667
#> [15,] 0.5000000 0.5000000 0.5000000 0.1666667 0.1666667 0.1666667
#> [16,] 0.5000000 0.5000000 0.5000000 0.1666667 0.1666667 0.1666667
#> [17,] 0.5000000 0.5000000 0.5000000 0.1666667 0.1666667 0.1666667
#> [18,] 0.1666667 0.1666667 0.1666667 0.5000000 0.5000000 0.5000000
#> [19,] 0.1666667 0.1666667 0.1666667 0.5000000 0.5000000 0.5000000
#> [20,] 0.1666667 0.1666667 0.1666667 0.5000000 0.5000000 0.5000000

#>   coefficient_name        est        se         p  ci_lower ci_upper df
#> 1                z -0.8231346 0.8539677 0.3633268 -2.792388 1.146118  8

# SE using collapsed cluster totals and Young's inequality
horvitz_thompson(y ~ z, data = dat, declaration = clust_crs_decl, collapsed = TRUE)

#>              y t clusters condition_probabilities
#> 1  -0.42938009 0        1                     0.5
#> 2   1.36046133 0        1                     0.5
#> 3  -0.07085743 0        2                     0.5
#> 4  -0.27215368 0        2                     0.5
#> 5  -2.44668003 1        3                     0.5
#> 6   0.06548664 1        3                     0.5
#> 7  -1.09850890 1        3                     0.5
#> 8  -0.63317818 1        4                     0.5
#> 9  -2.06365445 1        4                     0.5
#> 10  2.64893203 1        4                     0.5

#>   coefficient_name        est        se         p  ci_lower  ci_upper df
#> 1                z -0.8231346 0.7559133 0.3078984 -2.566274 0.9200045  8