Linear regression with the Lin (2013) covariate adjustment

This function is a wrapper for lm_robust that is useful for estimating treatment effects with pre-treatment covariate data. This implements the method described by Lin (2013).

lm_lin(formula, covariates, data, weights, subset, clusters, se_type = NULL,
  ci = TRUE, alpha = 0.05, return_vcov = TRUE, try_cholesky = FALSE)

Arguments

formula
an object of class formula, as in lm, such as Y ~ Z with only one variable on the right-hand side, the treatment

covariates
a right-sided formula with pre-treatment covaraites on the right hand side, such as ~ x1 + x2 + x3.

data
A data.frame

weights
the bare (unquoted) names of the weights variable in the supplied data.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

clusters
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.

se_type
The sort of standard error sought. If `clusters` is not specified the options are "HC0", "HC1" (or "stata", the equivalent), "HC2" (default), "HC3", or "classical". If `clusters` is specified the options are "CR0", "CR2" (default), or "stata" are permissible.

ci
logical. Whether to compute and return p-values and confidence intervals, TRUE by default.

alpha
The significance level, 0.05 by default.

return_vcov
logical. Whether to return the variance-covariance matrix for later usage, TRUE by default.

try_cholesky
logical. Whether to try using a Cholesky decomposition to solve least squares instead of a QR decomposition, FALSE by default. Using a Cholesky decomposition may result in speed gains, but should only be used if users are sure their model is full-rank (i.e., there is no perfect multi-collinearity)

Value

An object of class "lm_robust".

The post-estimation commands functions summary and tidy return results in a data.frame. To get useful data out of the return, you can use these data frames, you can use the resulting list directly, or you can use the generic accessor functions coef, vcov, confint, and predict. Marginal effects and uncertainty about them can be gotten by passing this object to margins from the margins.

Users who want to print the results in TeX of HTML can use the extract function and the texreg package.

An object of class "lm_robust" is a list containing at least the following components:

coefficients

the estimated coefficients

se

the estimated standard errors

df

the estimated degrees of freedom

p

the p-values from a two-sided t-test using coefficients, se, and df

ci_lower

the lower bound of the 1 - alpha percent confidence interval

ci_upper

the upper bound of the 1 - alpha percent confidence interval

coefficient_name

a character vector of coefficient names

alpha

the significance level specified by the user

res_var

the residual variance

N

the number of observations used

k

the number of columns in the design matrix (includes linearly dependent columns!)

rank

the rank of the fitted model

vcov

the fitted variance covariance matrix

r.squared

The \(R^2\), $$R^2 = 1 - Sum(e[i]^2) / Sum((y[i] - y^*)^2),$$ where \(y^*\) is the mean of \(y[i]\) if there is an intercept and zero otherwise, and \(e[i]\) is the ith residual.

adj.r.squared

The \(R^2\) but penalized for having more parameters, rank

weighted

whether or not weights were applied

call

the original function call

We also return terms and contrasts, used by predict, and scaled_centerthe means of each of the covariates used for centering them
Details

This function is simply a wrapper for lm_robust and implements the Lin estimator (see the reference below). This method pre-processes the data by taking the covariates specified in the `covariates` argument, centering them by subtracting from each covariate its mean, and interacting them with the treatment. If the treatment has multiple values, a series of dummies for each value is created and each of those is interacted with the demeaned covariates. More details can be found in the Getting Started vignette and the technical notes.

See also

lm_robust

#' @references Freedman, David A. 2008. "On Regression Adjustments in Experiments with Several Treatments." The Annals of Applied Statistics. JSTOR, 176-96. https://doi.org/10.1214/07-AOAS143.

Lin, Winston. 2013. "Agnostic Notes on Regression Adjustments to Experimental Data: Reexamining Freedman's Critique." The Annals of Applied Statistics 7 (1). Institute of Mathematical Statistics: 295-318. https://doi.org/10.1214/12-AOAS583.

Examples

library(fabricatr)
library(randomizr)
dat <- fabricate(
  N = 40,
  x = rnorm(N, mean = 2.3),
  x2 = rpois(N, lambda = 2),
  x3 = runif(N),
  y0 = rnorm(N) + x,
  y1 = rnorm(N) + x + 0.35
)

dat$z <- complete_ra(N = nrow(dat))
dat$y <- ifelse(dat$z == 1, dat$y1, dat$y0)

# Same specification as lm_robust() with one additional argument
lmlin_out <- lm_lin(y ~ z, covariates = ~ x, data = dat)
tidy(lmlin_out)

#>   coefficient_name coefficients        se            p   ci_lower  ci_upper df
#> 1      (Intercept)    1.8269388 0.2543331 1.890148e-08  1.3111275 2.3427502 36
#> 2                z    0.8431161 0.3380445 1.735734e-02  0.1575302 1.5287021 36
#> 3            x_bar    0.8616826 0.2580414 1.963654e-03  0.3383503 1.3850148 36
#> 4          z:x_bar    0.1255669 0.3445306 7.176483e-01 -0.5731736 0.8243074 36
#>   outcome
#> 1       y
#> 2       y
#> 3       y
#> 4       y

# Works with multiple pre-treatment covariates
lm_lin(y ~ z, covariates = ~ x + x2, data = dat)

#>                Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper DF
#> (Intercept)  1.79962363  0.2376886 8.557005e-09  1.3165823 2.2826650 34
#> z            0.92376708  0.3195841 6.655097e-03  0.2742941 1.5732401 34
#> x_bar        0.92380934  0.2770725 2.075764e-03  0.3607303 1.4868883 34
#> x2_bar      -0.12902605  0.1927653 5.077970e-01 -0.5207722 0.2627202 34
#> z:x_bar      0.07918477  0.3519194 8.233198e-01 -0.6360015 0.7943711 34
#> z:x2_bar    -0.13573956  0.2631839 6.093603e-01 -0.6705935 0.3991144 34

# Also centers data AFTER evaluating any functions in formula
lmlin_out2 <- lm_lin(y ~ z, covariates = ~ x + log(x3), data = dat)
lmlin_out2$scaled_center["log(x3)"]

#>    log(x3) 
#> -0.9732079 

mean(log(dat$x3))

#> [1] -0.9732079

# Works easily with clusters
dat$clusterID <- rep(1:20, each = 2)
dat$z_clust <- cluster_ra(clusters = dat$clusterID)

lm_lin(y ~ z_clust, covariates = ~ x, data = dat, clusters = clusterID)

#>                 Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper        DF
#> (Intercept)   2.06567229  0.1469872 4.202422e-07  1.7291873 2.4021572  8.351636
#> z_clust       0.36590730  0.3293214 2.821635e-01 -0.3294847 1.0612993 16.814419
#> x_bar         0.92669657  0.1452804 1.395332e-03  0.5533746 1.3000185  5.005936
#> z_clust:x_bar 0.02454839  0.2962328 9.358290e-01 -0.6489844 0.6980812  8.710723

# Works with multi-valued treatments
dat$z_multi <- sample(1:3, size = nrow(dat), replace = TRUE)
lm_lin(y ~ z_multi, covariates = ~ x, data = dat)

#>                   Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper DF
#> (Intercept)     2.45470820  0.2131209 2.775987e-13  2.0215943 2.8878221 34
#> z_multi2       -0.38865454  0.3778223 3.108974e-01 -1.1564818 0.3791727 34
#> z_multi3       -0.25508609  0.5184909 6.258965e-01 -1.3087863 0.7986141 34
#> x_bar           0.87410782  0.1339753 1.809979e-07  0.6018373 1.1463783 34
#> z_multi2:x_bar -0.02728243  0.3084666 9.300416e-01 -0.6541620 0.5995972 34
#> z_multi3:x_bar  0.15613744  0.4997146 7.566044e-01 -0.8594047 1.1716796 34

# Stratified estimator with blocks
dat$blockID <- rep(1:5, each = 8)
dat$z_block <- block_ra(blocks = dat$blockID)

lm_lin(y ~ z_block, ~ factor(blockID), data = dat)

#>                                 Estimate Std. Error     Pr(>|t|)   CI Lower
#> (Intercept)                   2.15370744  0.2826364 1.691145e-08  1.5764869
#> z_block                       0.18652698  0.4752187 6.974581e-01 -0.7839991
#> factor(blockID)2_bar         -0.93462363  1.0579586 3.840334e-01 -3.0952634
#> factor(blockID)3_bar         -1.21216762  0.7837649 1.324460e-01 -2.8128292
#> factor(blockID)4_bar         -0.03740077  0.7917444 9.626363e-01 -1.6543585
#> factor(blockID)5_bar         -0.16664690  0.6395678 7.962107e-01 -1.4728186
#> z_block:factor(blockID)2_bar  0.01735174  1.6629417 9.917438e-01 -3.3788283
#> z_block:factor(blockID)3_bar  1.74361108  1.5011288 2.545836e-01 -1.3221029
#> z_block:factor(blockID)4_bar  1.20744160  1.3967845 3.942038e-01 -1.6451729
#> z_block:factor(blockID)5_bar  0.20422731  1.3114802 8.772947e-01 -2.4741727
#>                               CI Upper DF
#> (Intercept)                  2.7309280 30
#> z_block                      1.1570531 30
#> factor(blockID)2_bar         1.2260161 30
#> factor(blockID)3_bar         0.3884939 30
#> factor(blockID)4_bar         1.5795569 30
#> factor(blockID)5_bar         1.1395248 30
#> z_block:factor(blockID)2_bar 3.4135318 30
#> z_block:factor(blockID)3_bar 4.8093251 30
#> z_block:factor(blockID)4_bar 4.0600561 30
#> z_block:factor(blockID)5_bar 2.8826273 30

# NOT RUN {
  # Can also use 'margins' package if you have it installed to get
  # marginal effects
  library(margins)
  lmlout <- lm_lin(y ~ z_block, ~ x, data = dat)
  summary(margins(lmlout))

  # Can output results using 'texreg'
  library(texreg)
  texregobj <- extract(lmlout)
# }