Linear regression with the Lin (2013) covariate adjustment

This function is a wrapper for lm_robust that is useful for estimating treatment effects with pre-treatment covariate data. This implements the method described by Lin (2013).

lm_lin(formula, covariates, data, weights, subset, clusters, se_type = NULL,
  ci = TRUE, alpha = 0.05, return_vcov = TRUE, try_cholesky = FALSE)

Arguments

formula
an object of class formula, as in lm, such as Y ~ Z with only one variable on the right-hand side, the treatment

covariates
a right-sided formula with pre-treatment covariates on the right hand side, such as ~ x1 + x2 + x3.

data
A data.frame

weights
the bare (unquoted) names of the weights variable in the supplied data.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

clusters
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.

se_type
The sort of standard error sought. If `clusters` is not specified the options are "HC0", "HC1" (or "stata", the equivalent), "HC2" (default), "HC3", or "classical". If `clusters` is specified the options are "CR0", "CR2" (default), or "stata" are permissible.

ci
logical. Whether to compute and return p-values and confidence intervals, TRUE by default.

alpha
The significance level, 0.05 by default.

return_vcov
logical. Whether to return the variance-covariance matrix for later usage, TRUE by default.

try_cholesky
logical. Whether to try using a Cholesky decomposition to solve least squares instead of a QR decomposition, FALSE by default. Using a Cholesky decomposition may result in speed gains, but should only be used if users are sure their model is full-rank (i.e., there is no perfect multi-collinearity)

Value

An object of class "lm_robust".

The post-estimation commands functions summary and tidy return results in a data.frame. To get useful data out of the return, you can use these data frames, you can use the resulting list directly, or you can use the generic accessor functions coef, vcov, confint, and predict. Marginal effects and uncertainty about them can be gotten by passing this object to margins from the margins.

Users who want to print the results in TeX of HTML can use the extract function and the texreg package.

An object of class "lm_robust" is a list containing at least the following components:

coefficients

the estimated coefficients

std.error

the estimated standard errors

df

the estimated degrees of freedom

p.value

the p-values from a two-sided t-test using coefficients, std.error, and df

ci.lower

the lower bound of the 1 - alpha percent confidence interval

ci.upper

the upper bound of the 1 - alpha percent confidence interval

term

a character vector of coefficient names

alpha

the significance level specified by the user

se_type

the standard error type specified by the user

res_var

the residual variance

N

the number of observations used

k

the number of columns in the design matrix (includes linearly dependent columns!)

rank

the rank of the fitted model

vcov

the fitted variance covariance matrix

r.squared

The \(R^2\), $$R^2 = 1 - Sum(e[i]^2) / Sum((y[i] - y^*)^2),$$ where \(y^*\) is the mean of \(y[i]\) if there is an intercept and zero otherwise, and \(e[i]\) is the ith residual.

adj.r.squared

The \(R^2\) but penalized for having more parameters, rank

weighted

whether or not weights were applied

call

the original function call

We also return terms and contrasts, used by predict, and scaled_centerthe means of each of the covariates used for centering them
Details

This function is simply a wrapper for lm_robust and implements the Lin estimator (see the reference below). This method pre-processes the data by taking the covariates specified in the `covariates` argument, centering them by subtracting from each covariate its mean, and interacting them with the treatment. If the treatment has multiple values, a series of dummies for each value is created and each of those is interacted with the demeaned covariates. More details can be found in the Getting Started vignette and the mathematical notes.

See also

lm_robust

#' @references Freedman, David A. 2008. "On Regression Adjustments in Experiments with Several Treatments." The Annals of Applied Statistics. JSTOR, 176-96. https://doi.org/10.1214/07-AOAS143.

Lin, Winston. 2013. "Agnostic Notes on Regression Adjustments to Experimental Data: Reexamining Freedman's Critique." The Annals of Applied Statistics 7 (1). Institute of Mathematical Statistics: 295-318. https://doi.org/10.1214/12-AOAS583.

Examples

library(fabricatr)
library(randomizr)
dat <- fabricate(
  N = 40,
  x = rnorm(N, mean = 2.3),
  x2 = rpois(N, lambda = 2),
  x3 = runif(N),
  y0 = rnorm(N) + x,
  y1 = rnorm(N) + x + 0.35
)

dat$z <- complete_ra(N = nrow(dat))
dat$y <- ifelse(dat$z == 1, dat$y1, dat$y0)

# Same specification as lm_robust() with one additional argument
lmlin_out <- lm_lin(y ~ z, covariates = ~ x, data = dat)
tidy(lmlin_out)

#>          term   estimate std.error      p.value   ci.lower  ci.upper df outcome
#> 1 (Intercept)  2.5472789 0.2490300 3.382685e-12  2.0422227 3.0523351 36       y
#> 2           z -0.2615416 0.3240810 4.249518e-01 -0.9188083 0.3957251 36       y
#> 3         x_c  1.0101617 0.3880989 1.334249e-02  0.2230605 1.7972628 36       y
#> 4       z:x_c  0.3701714 0.4535144 4.197378e-01 -0.5495984 1.2899413 36       y

# Works with multiple pre-treatment covariates
lm_lin(y ~ z, covariates = ~ x + x2, data = dat)

#>               Estimate Std. Error     Pr(>|t|)    CI Lower     CI Upper DF
#> (Intercept)  2.4345143  0.2366213 5.600646e-12  1.95364190  2.915386674 34
#> z           -0.1229709  0.3117459 6.957060e-01 -0.75651480  0.510573057 34
#> x_c          0.7923640  0.3761990 4.264235e-02  0.02783567  1.556892291 34
#> x2_c        -0.2568127  0.1250432 4.775041e-02 -0.51093111 -0.002694208 34
#> z:x_c        0.5690671  0.4645692 2.290185e-01 -0.37505097  1.513185245 34
#> z:x2_c       0.1544267  0.1710863 3.730756e-01 -0.19326241  0.502115789 34

# Also centers data AFTER evaluating any functions in formula
lmlin_out2 <- lm_lin(y ~ z, covariates = ~ x + log(x3), data = dat)
lmlin_out2$scaled_center["log(x3)"]

#>   log(x3) 
#> -1.147778 

mean(log(dat$x3))

#> [1] -1.147778

# Works easily with clusters
dat$clusterID <- rep(1:20, each = 2)
dat$z_clust <- cluster_ra(clusters = dat$clusterID)

lm_lin(y ~ z_clust, covariates = ~ x, data = dat, clusters = clusterID)

#>               Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper        DF
#> (Intercept)  2.5621501  0.2110458 1.624773e-06  2.0773167 3.0469835  8.178139
#> z_clust     -0.3278225  0.3568058 3.718005e-01 -1.0839465 0.4283015 16.070708
#> x_c          1.4145819  0.2286993 7.783283e-04  0.8568906 1.9722733  6.086205
#> z_clust:x_c -0.6918015  0.4975998 1.919935e-01 -1.7873087 0.4037057 10.975570

# Works with multi-valued treatments
dat$z_multi <- sample(1:3, size = nrow(dat), replace = TRUE)
lm_lin(y ~ z_multi, covariates = ~ x, data = dat)

#>               Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper DF
#> (Intercept)  2.3734578  0.3344897 3.380265e-08  1.6936929 3.0532226 34
#> z_multi2     0.1854411  0.3725737 6.218784e-01 -0.5717197 0.9426019 34
#> z_multi3     0.1321559  0.4100717 7.492176e-01 -0.7012101 0.9655219 34
#> x_c          0.7367735  0.5355331 1.778871e-01 -0.3515608 1.8251077 34
#> z_multi2:x_c 0.7728239  0.5672827 1.820523e-01 -0.3800331 1.9256810 34
#> z_multi3:x_c 0.2992611  0.6884900 6.665535e-01 -1.0999190 1.6984411 34

# Stratified estimator with blocks
dat$blockID <- rep(1:5, each = 8)
dat$z_block <- block_ra(blocks = dat$blockID)

lm_lin(y ~ z_block, ~ factor(blockID), data = dat)

#>                                Estimate Std. Error     Pr(>|t|)   CI Lower
#> (Intercept)                 2.466490923  0.3696351 2.163152e-07  1.7115953
#> z_block                     0.004394078  0.4304694 9.919232e-01 -0.8747418
#> factor(blockID)2_c          0.527245243  1.4226208 7.135281e-01 -2.3781341
#> factor(blockID)3_c          0.113398391  1.4642100 9.387823e-01 -2.8769173
#> factor(blockID)4_c         -0.348750385  1.3223663 7.937896e-01 -3.0493826
#> factor(blockID)5_c          1.110661232  1.2410357 3.779400e-01 -1.4238719
#> z_block:factor(blockID)2_c -1.530559180  1.5456602 3.299817e-01 -4.6872184
#> z_block:factor(blockID)3_c -0.387863686  1.5607355 8.054306e-01 -3.5753108
#> z_block:factor(blockID)4_c  0.503594185  1.6003540 7.551853e-01 -2.7647647
#> z_block:factor(blockID)5_c -0.864156502  1.3671117 5.321071e-01 -3.6561711
#>                             CI Upper DF
#> (Intercept)                3.2213866 30
#> z_block                    0.8835299 30
#> factor(blockID)2_c         3.4326246 30
#> factor(blockID)3_c         3.1037141 30
#> factor(blockID)4_c         2.3518819 30
#> factor(blockID)5_c         3.6451943 30
#> z_block:factor(blockID)2_c 1.6261001 30
#> z_block:factor(blockID)3_c 2.7995834 30
#> z_block:factor(blockID)4_c 3.7719531 30
#> z_block:factor(blockID)5_c 1.9278581 30

# NOT RUN {
  # Can also use 'margins' package if you have it installed to get
  # marginal effects
  library(margins)
  lmlout <- lm_lin(y ~ z_block, ~ x, data = dat)
  summary(margins(lmlout))

  # Can output results using 'texreg'
  library(texreg)
  texregobj <- extract(lmlout)
# }