Linear regression with the Lin (2013) covariate adjustment

This function is a wrapper for lm_robust that is useful for estimating treatment effects with pre-treatment covariate data. This implements the method described by Lin (2013).

lm_lin(formula, covariates, data, weights, subset, clusters, se_type = NULL,
  ci = TRUE, alpha = 0.05, return_vcov = TRUE, try_cholesky = FALSE)

Arguments

formula
an object of class formula, as in lm, such as Y ~ Z with only one variable on the right-hand side, the treatment

covariates
a right-sided formula with pre-treatment covariates on the right hand side, such as ~ x1 + x2 + x3.

data
A data.frame

weights
the bare (unquoted) names of the weights variable in the supplied data.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

clusters
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.

se_type
The sort of standard error sought. If `clusters` is not specified the options are "HC0", "HC1" (or "stata", the equivalent), "HC2" (default), "HC3", or "classical". If `clusters` is specified the options are "CR0", "CR2" (default), or "stata" are permissible.

ci
logical. Whether to compute and return p-values and confidence intervals, TRUE by default.

alpha
The significance level, 0.05 by default.

return_vcov
logical. Whether to return the variance-covariance matrix for later usage, TRUE by default.

try_cholesky
logical. Whether to try using a Cholesky decomposition to solve least squares instead of a QR decomposition, FALSE by default. Using a Cholesky decomposition may result in speed gains, but should only be used if users are sure their model is full-rank (i.e., there is no perfect multi-collinearity)

Value

An object of class "lm_robust".

The post-estimation commands functions summary and tidy return results in a data.frame. To get useful data out of the return, you can use these data frames, you can use the resulting list directly, or you can use the generic accessor functions coef, vcov, confint, and predict. Marginal effects and uncertainty about them can be gotten by passing this object to margins from the margins.

Users who want to print the results in TeX of HTML can use the extract function and the texreg package.

An object of class "lm_robust" is a list containing at least the following components:

coefficients

the estimated coefficients

std.error

the estimated standard errors

statistic

the t-statistic

df

the estimated degrees of freedom

p.value

the p-values from a two-sided t-test using coefficients, std.error, and df

conf.low

the lower bound of the 1 - alpha percent confidence interval

conf.high

the upper bound of the 1 - alpha percent confidence interval

term

a character vector of coefficient names

alpha

the significance level specified by the user

se_type

the standard error type specified by the user

res_var

the residual variance

N

the number of observations used

k

the number of columns in the design matrix (includes linearly dependent columns!)

rank

the rank of the fitted model

vcov

the fitted variance covariance matrix

r.squared

The \(R^2\), $$R^2 = 1 - Sum(e[i]^2) / Sum((y[i] - y^*)^2),$$ where \(y^*\) is the mean of \(y[i]\) if there is an intercept and zero otherwise, and \(e[i]\) is the ith residual.

adj.r.squared

The \(R^2\) but penalized for having more parameters, rank

weighted

whether or not weights were applied

call

the original function call

fitted.values

the matrix of predicted means

We also return terms and contrasts, used by predict, and scaled_centerthe means of each of the covariates used for centering them
Details

This function is simply a wrapper for lm_robust and implements the Lin estimator (see the reference below). This method pre-processes the data by taking the covariates specified in the `covariates` argument, centering them by subtracting from each covariate its mean, and interacting them with the treatment. If the treatment has multiple values, a series of dummies for each value is created and each of those is interacted with the demeaned covariates. More details can be found in the Getting Started vignette and the mathematical notes.

See also

lm_robust

#' @references Freedman, David A. 2008. "On Regression Adjustments in Experiments with Several Treatments." The Annals of Applied Statistics. JSTOR, 176-96. https://doi.org/10.1214/07-AOAS143.

Lin, Winston. 2013. "Agnostic Notes on Regression Adjustments to Experimental Data: Reexamining Freedman's Critique." The Annals of Applied Statistics 7 (1). Institute of Mathematical Statistics: 295-318. https://doi.org/10.1214/12-AOAS583.

Examples

library(fabricatr)
library(randomizr)
dat <- fabricate(
  N = 40,
  x = rnorm(N, mean = 2.3),
  x2 = rpois(N, lambda = 2),
  x3 = runif(N),
  y0 = rnorm(N) + x,
  y1 = rnorm(N) + x + 0.35
)

dat$z <- complete_ra(N = nrow(dat))
dat$y <- ifelse(dat$z == 1, dat$y1, dat$y0)

# Same specification as lm_robust() with one additional argument
lmlin_out <- lm_lin(y ~ z, covariates = ~ x, data = dat)
tidy(lmlin_out)

#>          term  estimate std.error statistic      p.value   conf.low conf.high
#> 1 (Intercept) 1.8026178 0.2165801  8.323100 6.569833e-10  1.3633730  2.241863
#> 2           z 1.0316029 0.3142429  3.282821 2.291303e-03  0.3942889  1.668917
#> 3         x_c 0.5777179 0.2290022  2.522762 1.620364e-02  0.1132800  1.042156
#> 4       z:x_c 0.4791451 0.2927459  1.636727 1.104013e-01 -0.1145712  1.072861
#>   df outcome
#> 1 36       y
#> 2 36       y
#> 3 36       y
#> 4 36       y

# Works with multiple pre-treatment covariates
lm_lin(y ~ z, covariates = ~ x + x2, data = dat)

#>               Estimate Std. Error   t value     Pr(>|t|)    CI Lower
#> (Intercept)  1.8259819 0.20588206  8.869067 2.296005e-10  1.40757918
#> z            1.0077530 0.30932172  3.257944 2.548062e-03  0.37913562
#> x_c          0.6147020 0.22697321  2.708258 1.051063e-02  0.15343689
#> x2_c        -0.2372305 0.08687123 -2.730829 9.939845e-03 -0.41377409
#> z:x_c        0.4343617 0.29518481  1.471491 1.503545e-01 -0.16552603
#> z:x2_c       0.2150855 0.14829333  1.450406 1.561089e-01 -0.08628277
#>                CI Upper DF
#> (Intercept)  2.24438456 34
#> z            1.63637034 34
#> x_c          1.07596702 34
#> x2_c        -0.06068691 34
#> z:x_c        1.03424941 34
#> z:x2_c       0.51645385 34

# Also centers data AFTER evaluating any functions in formula
lmlin_out2 <- lm_lin(y ~ z, covariates = ~ x + log(x3), data = dat)
lmlin_out2$scaled_center["log(x3)"]

#>  log(x3) 
#> -1.05676 

mean(log(dat$x3))

#> [1] -1.05676

# Works easily with clusters
dat$clusterID <- rep(1:20, each = 2)
dat$z_clust <- cluster_ra(clusters = dat$clusterID)

lm_lin(y ~ z_clust, covariates = ~ x, data = dat, clusters = clusterID)

#>               Estimate Std. Error    t value     Pr(>|t|)   CI Lower  CI Upper
#> (Intercept)  2.2858578  0.2499567  9.1450135 1.017325e-05  1.7164169 2.8552986
#> z_clust      0.1645257  0.3942542  0.4173085 6.816056e-01 -0.6664243 0.9954756
#> x_c          1.0041073  0.1938795  5.1790277 4.112009e-03  0.4974237 1.5107910
#> z_clust:x_c -0.1976663  0.3457910 -0.5716352 5.790456e-01 -0.9586096 0.5632771
#>                    DF
#> (Intercept)  8.603305
#> z_clust     17.232425
#> x_c          4.741030
#> z_clust:x_c 11.016320

# Works with multi-valued treatments
dat$z_multi <- sample(1:3, size = nrow(dat), replace = TRUE)
lm_lin(y ~ z_multi, covariates = ~ x, data = dat)

#>                 Estimate Std. Error     t value     Pr(>|t|)   CI Lower
#> (Intercept)   2.44175655  0.3181705  7.67436549 6.374489e-09  1.7951563
#> z_multi2      0.03881805  0.4732673  0.08202140 9.351107e-01 -0.9229768
#> z_multi3     -0.14277905  0.4336806 -0.32922626 7.440042e-01 -1.0241241
#> x_c           1.10254997  0.3469138  3.17816688 3.151395e-03  0.3975363
#> z_multi2:x_c  0.03000709  0.5223335  0.05744814 9.545245e-01 -1.0315022
#> z_multi3:x_c -0.36055874  0.4378240 -0.82352440 4.159492e-01 -1.2503242
#>               CI Upper DF
#> (Intercept)  3.0883568 34
#> z_multi2     1.0006129 34
#> z_multi3     0.7385660 34
#> x_c          1.8075636 34
#> z_multi2:x_c 1.0915164 34
#> z_multi3:x_c 0.5292067 34

# Stratified estimator with blocks
dat$blockID <- rep(1:5, each = 8)
dat$z_block <- block_ra(blocks = dat$blockID)

lm_lin(y ~ z_block, ~ factor(blockID), data = dat)

#>                              Estimate Std. Error    t value     Pr(>|t|)
#> (Intercept)                 2.1666198  0.3125362  6.9323807 1.065456e-07
#> z_block                     0.3757844  0.4980395  0.7545274 4.564151e-01
#> factor(blockID)2_c          0.5941769  0.8270890  0.7183953 4.780717e-01
#> factor(blockID)3_c          0.5165867  1.2124479  0.4260692 6.730990e-01
#> factor(blockID)4_c          1.2301622  0.8766283  1.4032883 1.707971e-01
#> factor(blockID)5_c          1.5991702  0.7227921  2.2124899 3.468740e-02
#> z_block:factor(blockID)2_c  0.4296514  1.6620499  0.2585069 7.977811e-01
#> z_block:factor(blockID)3_c -1.3078623  1.6407573 -0.7971089 4.316516e-01
#> z_block:factor(blockID)4_c -0.2990054  1.4411165 -0.2074818 8.370357e-01
#> z_block:factor(blockID)5_c -0.5548165  1.5465141 -0.3587530 7.222924e-01
#>                              CI Lower CI Upper DF
#> (Intercept)                 1.5283357 2.804904 30
#> z_block                    -0.6413478 1.392917 30
#> factor(blockID)2_c         -1.0949642 2.283318 30
#> factor(blockID)3_c         -1.9595622 2.992736 30
#> factor(blockID)4_c         -0.5601515 3.020476 30
#> factor(blockID)5_c          0.1230318 3.075309 30
#> z_block:factor(blockID)2_c -2.9647073 3.824010 30
#> z_block:factor(blockID)3_c -4.6587357 2.043011 30
#> z_block:factor(blockID)4_c -3.2421579 2.644147 30
#> z_block:factor(blockID)5_c -3.7132197 2.603587 30

# NOT RUN {
  # Can also use 'margins' package if you have it installed to get
  # marginal effects
  library(margins)
  lmlout <- lm_lin(y ~ z_block, ~ x, data = dat)
  summary(margins(lmlout))

  # Can output results using 'texreg'
  library(texreg)
  texregobj <- extract(lmlout)
# }