Linear regression with the Lin (2013) covariate adjustment

This function is a wrapper for lm_robust that is useful for estimating treatment effects with pre-treatment covariate data. This implements the method described by Lin (2013) to reduce the bias of such estimation

lm_lin(formula, covariates, data, weights, subset, clusters, se_type = NULL,
  ci = TRUE, alpha = 0.05, return_vcov = TRUE, try_cholesky = FALSE)

Arguments

formula
an object of class formula, as in lm, such as Y ~ Z with only one variable on the right-hand side, the treatment

covariates
a right-sided formula with pre-treatment covaraites on the right hand side, such as ~ x1 + x2 + x3.

data
A data.frame

weights
the bare (unquoted) names of the weights variable in the supplied data.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

clusters
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.

se_type
The sort of standard error sought. Without clustering: "HC0", "HC1" (or "stata", the equivalent), "HC2" (default), "HC3", or "classical". With clustering: "CR0", "CR2" (default), or "stata" are permissible.

ci
A boolean for whether to compute and return p-values and confidence intervals, TRUE by default.

alpha
The significance level, 0.05 by default.

return_vcov
a boolean for whether to return the variance-covariance matrix for later usage, TRUE by default.

try_cholesky
a boolean for whether to try using a Cholesky decomposition to solve LS instead of a QR decomposition, FALSE by default. Using a Cholesky decomposition may result in speed gains, but should only be used if users are sure their model is full-rank (i.e. there is no perfect multi-collinearity).

Value

lm_lin returns an object of class "lm_robust".

The functions summary and tidy can be used to get the results as a data.frame. To get useful data out of the return, you can use these data frames, you can use the resulting list directly, or you can use the generic accessor functions coef, vcov, confint, and predict.

An object of class "lm_robust" is a list containing at least the following components:

est

the estimated coefficients

se

the estimated standard errors

df

the estimated degrees of freedom

p

the p-values from the t-test using est, se, and df

ci_lower

the lower bound of the 1 - alpha percent confidence interval

ci_upper

the upper bound of the 1 - alpha percent confidence interval

coefficient_name

a character vector of coefficient names

alpha

the significance level specified by the user

res_var

the residual variance, used for uncertainty when using predict

N

the number of observations used

k

the number of columns in the design matrix (includes linearly dependent columns!)

rank

the rank of the fitted model

vcov

the fitted variance covariance matrix

weighted

whether or not weights were applied

scaled_center

the means of each of the covariates used for centering them

We also return terms and contrasts, used by predict.

Details

This function is simply a wrapper for lm_robust. This method pre-processes the data by taking the covariates specified in the `covariates` argument, centering them by subtracting from each covariate its mean, and interacting them with the treatment. If the treatment has multiple values, a series of dummies for each value is created and each of those is interacted with the demeaned covariates. More details can be found in the Getting Started vignette and the technical notes.

References

Freedman, David A. 2008. "On Regression Adjustments in Experiments with Several Treatments." The Annals of Applied Statistics. JSTOR, 176-96. https://doi.org/10.1214/07-AOAS143. Lin, Winston. 2013. "Agnostic Notes on Regression Adjustments to Experimental Data: Reexamining Freedman's Critique." The Annals of Applied Statistics 7 (1). Institute of Mathematical Statistics: 295-318. https://doi.org/10.1214/12-AOAS583.

See also

lm_robust

Examples

library(fabricatr)
library(randomizr)
dat <- fabricate(
  N = 40,
  x = rnorm(N, mean = 2.3),
  x2 = rpois(N, lambda = 2),
  x3 = runif(N),
  y0 = rnorm(N) + x,
  y1 = rnorm(N) + x + 0.35
)

dat$z <- simple_ra(N = nrow(dat))
dat$y <- ifelse(dat$z == 1, dat$y1, dat$y0)

# Same specification as `lm_robust()` with one additional argument
lmlin_out <- lm_lin(y ~ z, covariates = ~ x, data = dat)
tidy(lmlin_out)

#>   coefficient_name       est        se            p   ci_lower  ci_upper df
#> 1      (Intercept) 2.1475300 0.2618931 9.377889e-10  1.6163862 2.6786738 36
#> 2                z 0.3077413 0.3411292 3.729868e-01 -0.3841008 0.9995834 36
#> 3            x_bar 0.9654595 0.2206322 9.929515e-05  0.5179966 1.4129224 36
#> 4          z:x_bar 0.1292807 0.3346727 7.015556e-01 -0.5494670 0.8080284 36
#>   outcome
#> 1       y
#> 2       y
#> 3       y
#> 4       y

# Works with multiple pre-treatment covariates
lm_lin(y ~ z, covariates = ~ x + x2, data = dat)

#>   coefficient_name         est        se            p   ci_lower  ci_upper df
#> 1      (Intercept)  2.17068526 0.2705310 2.368917e-09  1.6209000 2.7204705 34
#> 2                z  0.26577482 0.3476341 4.498286e-01 -0.4407027 0.9722523 34
#> 3            x_bar  1.03176130 0.1941414 6.707867e-06  0.6372185 1.4263040 34
#> 4           x2_bar -0.20659553 0.2389648 3.933480e-01 -0.6922304 0.2790393 34
#> 5          z:x_bar  0.08278470 0.3118637 7.922646e-01 -0.5509987 0.7165681 34
#> 6         z:x2_bar  0.06558611 0.2731794 8.117065e-01 -0.4895811 0.6207534 34
#>   outcome
#> 1       y
#> 2       y
#> 3       y
#> 4       y
#> 5       y
#> 6       y

# Also centers data AFTER evaluating any functions in formula
lmlin_out2 <- lm_lin(y ~ z, covariates = ~ x + log(x3), data = dat)
lmlin_out2$scaled_center["log(x3)"]

#>    log(x3) 
#> -0.9732079 

mean(log(dat$x3))

#> [1] -0.9732079

# Works easily with clusters
dat$clusterID <- rep(1:20, each = 2)
dat$z_clust <- cluster_ra(clusters = dat$clusterID)

lm_lin(y ~ z_clust, covariates = ~ x, data = dat, clusters = clusterID)

#>   coefficient_name        est        se            p   ci_lower  ci_upper
#> 1      (Intercept) 2.28937959 0.2450574 2.299263e-05  1.7165491 2.8622101
#> 2          z_clust 0.06334119 0.3547253 8.605812e-01 -0.6902927 0.8169751
#> 3            x_bar 0.87149505 0.1797028 5.324095e-03  0.4024771 1.3405130
#> 4    z_clust:x_bar 0.23905610 0.2672557 3.948223e-01 -0.3674290 0.8455412
#>          df outcome
#> 1  7.423751       y
#> 2 15.580257       y
#> 3  4.760539       y
#> 4  8.817709       y

# Works with multi-valued treatments
dat$z_multi <- sample(1:3, size = nrow(dat), replace = TRUE)
lm_lin(y ~ z_multi, covariates = ~ x, data = dat)

#>   coefficient_name        est        se            p   ci_lower  ci_upper df
#> 1      (Intercept)  2.4901921 0.2029913 4.876519e-14  2.0776642 2.9027199 34
#> 2         z_multi2 -0.4751601 0.3458558 1.784735e-01 -1.1780236 0.2277035 34
#> 3         z_multi3  0.1126915 0.4781834 8.151058e-01 -0.8590941 1.0844771 34
#> 4            x_bar  0.7706609 0.4792864 1.170977e-01 -0.2033662 1.7446880 34
#> 5   z_multi2:x_bar  0.2616373 0.5400056 6.311318e-01 -0.8357862 1.3590607 34
#> 6   z_multi3:x_bar  0.2105992 0.6112720 7.325726e-01 -1.0316549 1.4528534 34
#>   outcome
#> 1       y
#> 2       y
#> 3       y
#> 4       y
#> 5       y
#> 6       y