Ordinary Least Squares with Robust Standard Errors

This formula fits a linear model, provides a variety of options for robust standard errors, and conducts coefficient tests

lm_robust(formula, data, weights, subset, clusters, se_type = NULL,
  ci = TRUE, alpha = 0.05, return_vcov = TRUE, try_cholesky = FALSE)

Arguments

formula
an object of class formula, as in lm

data
A data.frame

weights
the bare (unquoted) names of the weights variable in the supplied data.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

clusters
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.

se_type
The sort of standard error sought. Without clustering: "HC0", "HC1" (or "stata", the equivalent), "HC2" (default), "HC3", or "classical". With clustering: "CR0", "CR2" (default), or "stata". are permissible.

ci
A boolean for whether to compute and return p-values and confidence intervals, TRUE by default.

alpha
The significance level, 0.05 by default.

return_vcov
a boolean for whether to return the variance-covariance matrix for later usage, TRUE by default.

try_cholesky
a boolean for whether to try using a Cholesky decomposition to solve LS instead of a QR decomposition, FALSE by default. Using a Cholesky decomposition may result in speed gains, but should only be used if users are sure their model is full-rank (i.e. there is no perfect multi-collinearity)

Value

lm_robust returns an object of class "lm_robust".

The functions summary and tidy can be used to get the results as a data.frame. To get useful data out of the return, you can use these data frames, you can use the resulting list directly, or you can use the generic accessor functions coef, vcov, confint, and predict.

An object of class "lm_robust" is a list containing at least the following components:

est

the estimated coefficients

se

the estimated standard errors

df

the estimated degrees of freedom

p

the p-values from the t-test using est, se, and df

ci_lower

the lower bound of the 1 - alpha percent confidence interval

ci_upper

the upper bound of the 1 - alpha percent confidence interval

coefficient_name

a character vector of coefficient names

alpha

the significance level specified by the user

res_var

the residual variance, used for uncertainty when using predict

N

the number of observations used

k

the number of columns in the design matrix (includes linearly dependent columns!)

rank

the rank of the fitted model

vcov

the fitted variance covariance matrix

weighted

whether or not weights were applied

We also return terms and contrasts, used by predict.

Details

This function does linear regression and provides a variety of standard errors. It takes formulae and data much in the same was as lm does, and all auxiliary variables, such as clusters and weights, can be passed either as quoted names of columns, as bare column names, or as a self-contained vector. Examples of usage can be seen below and in the Getting Started vignette.

To see what the exact equations and computations each different type of standard error entails, users can see the technical notes in this vignette. The default estimators for the non-clustered and clustered case have been chosen for desirable similarities to randomization estimators, performance in small samples, or some combination of both. The default for the case without clusters is the HC2 estimator and the default with clusters is the analogous CR2 estimator. Users can easily replicate Stata standard errors in the clustered or non-clustered case by setting se_type = "stata".

The function estimates the coefficients and standard errors in C++, using the RcppEigen package. By default, we estimate the coefficients using Column-Pivoting QR decomposition from the Eigen C++ library, although users could get faster solutions by setting try_cholesky = TRUE to use a Cholesky decomposition instead. This will likely result in quicker solutions, but the algorithm does not reliably detect when there are linear dependencies in the model and may fail silently if they exist.

References

Abadie, Alberto, Susan Athey, Guido W Imbens, and Jeffrey Wooldridge. 2017. "A Class of Unbiased Estimators of the Average Treatment Effect in Randomized Experiments." arXiv Pre-Print. https://arxiv.org/abs/1710.02926v2. Bell, Robert M, and Daniel F McCaffrey. 2002. "Bias Reduction in Standard Errors for Linear Regression with Multi-Stage Samples." Survey Methodology 28 (2): 169-82. MacKinnon, James, and Halbert White. 1985. "Some Heteroskedasticity-Consistent Covariance Matrix Estimators with Improved Finite Sample Properties." Journal of Econometrics 29 (3): 305-25. https://doi.org/10.1016/0304-4076(85)90158-7. Pustejovsky, James E, and Elizabeth Tipton. 2016. "Small Sample Methods for Cluster-Robust Variance Estimation and Hypothesis Testing in Fixed Effects Models." Journal of Business & Economic Statistics. Taylor & Francis. https://doi.org/10.1080/07350015.2016.1247004. Samii, Cyrus, and Peter M Aronow. 2012. "On Equivalencies Between Design-Based and Regression-Based Variance Estimators for Randomized Experiments." Statistics and Probability Letters 82 (2). https://doi.org/10.1016/j.spl.2011.10.024.

Examples

library(fabricatr)
dat <- fabricate(
  N = 40,
  y = rpois(N, lambda = 4),
  x = rnorm(N),
  z = rbinom(N, 1, prob = 0.4)
)

# Default is to use HC2 robust standard errors
lmro <- lm_robust(y ~ x + z, data = dat)

# Can `tidy()` or `summary()` the data to easily get output as a data.frame
tidy(lmro)

#>   coefficient_name       est        se            p   ci_lower ci_upper df
#> 1      (Intercept) 3.8270190 0.4251936 7.460320e-11  2.9654951 4.688543 37
#> 2                x 0.5218375 0.3070137 9.757698e-02 -0.1002313 1.143906 37
#> 3                z 1.0450491 0.7799145 1.884321e-01 -0.5352077 2.625306 37
#>   outcome
#> 1       y
#> 2       y
#> 3       y

summary(lmro)

#> $coefficients
#>              Estimate Std. Error     Pr(>|t|)   ci_lower ci_upper df
#> (Intercept) 3.8270190  0.4251936 7.460320e-11  2.9654951 4.688543 37
#> x           0.5218375  0.3070137 9.757698e-02 -0.1002313 1.143906 37
#> z           1.0450491  0.7799145 1.884321e-01 -0.5352077 2.625306 37
#> 

# Can also get coefficients many ways
all.equal(
  lmro$est,
  coef(lmro),
  tidy(lmro)$est,
  check.attributes = FALSE
)

#> [1] TRUE

# Can recover classical standard errors
lmclassic <- lm_robust(y ~ x + z, data = dat, se_type = "classical")
tidy(lmclassic)

#>   coefficient_name       est        se            p   ci_lower ci_upper df
#> 1      (Intercept) 3.8270190 0.4265135 8.078071e-11  2.9628206 4.691217 37
#> 2                x 0.5218375 0.3374516 1.305178e-01 -0.1619043 1.205579 37
#> 3                z 1.0450491 0.7434457 1.681603e-01 -0.4613150 2.551413 37
#>   outcome
#> 1       y
#> 2       y
#> 3       y

# Can easily match Stata's robust standard errors
lmstata <- lm_robust(y ~ x + z, data = dat, se_type = "stata")
tidy(lmstata)

#>   coefficient_name       est        se            p    ci_lower ci_upper df
#> 1      (Intercept) 3.8270190 0.4244077 7.113848e-11  2.96708743 4.686951 37
#> 2                x 0.5218375 0.3006726 9.096321e-02 -0.08738304 1.131058 37
#> 3                z 1.0450491 0.7728504 1.845187e-01 -0.52089458 2.610993 37
#>   outcome
#> 1       y
#> 2       y
#> 3       y

# Easy to specify clusters for cluster-robust inference
dat$clusterID <- sample(1:10, size = 40, replace = TRUE)

lmclust <- lm_robust(y ~ x + z, data = dat, clusters = clusterID)
tidy(lmclust)

#>   coefficient_name       est        se            p   ci_lower ci_upper
#> 1      (Intercept) 3.8270190 0.5179284 0.0001989159  2.5876424 5.066396
#> 2                x 0.5218375 0.3331261 0.1841796068 -0.3628153 1.406490
#> 3                z 1.0450491 0.7274699 0.2145028268 -0.8675556 2.957654
#>         df outcome
#> 1 6.613231       y
#> 2 4.516445       y
#> 3 4.654117       y

# Can also match Stata's clustered standard errors
lm_robust(
  y ~ x + z,
  data = dat,
  clusters = clusterID,
  se_type = "stata"
)

#>   coefficient_name       est        se            p   ci_lower ci_upper df
#> 1      (Intercept) 3.8270190 0.5062160 6.545973e-05  2.6596829 4.994355  8
#> 2                x 0.5218375 0.3181514 1.395898e-01 -0.2118208 1.255496  8
#> 3                z 1.0450491 0.7148910 1.819272e-01 -0.6034924 2.693591  8
#>   outcome
#> 1       y
#> 2       y
#> 3       y

# Works just as LM does with functions in the formula
dat$blockID <- rep(c("A", "B", "C", "D"), each = 10)

lm_robust(y ~ x + z + factor(blockID), data = dat)

#>   coefficient_name        est        se            p    ci_lower ci_upper df
#> 1      (Intercept)  3.6054786 0.7517428 3.151803e-05  2.07775335 5.133204 34
#> 2                x  0.4637445 0.2704874 9.554453e-02 -0.08595212 1.013441 34
#> 3                z  1.1270102 0.7640929 1.494227e-01 -0.42581336 2.679834 34
#> 4 factor(blockID)B  0.8725286 0.9039197 3.412245e-01 -0.96445717 2.709514 34
#> 5 factor(blockID)C  0.1950133 1.0867954 8.586583e-01 -2.01362069 2.403647 34
#> 6 factor(blockID)D -0.3534448 0.9467730 7.112314e-01 -2.27751896 1.570629 34
#>   outcome
#> 1       y
#> 2       y
#> 3       y
#> 4       y
#> 5       y
#> 6       y

# Weights are also easily specified
dat$w <- runif(40)

lm_robust(
  y ~ x + z,
  data = dat,
  weights = w,
  clusters = clusterID
)

#>   coefficient_name      est        se            p   ci_lower ci_upper       df
#> 1      (Intercept) 3.378684 0.4139408 0.0001004981  2.3913906 4.365977 6.715198
#> 2                x 0.391359 0.3201016 0.2941650911 -0.5299752 1.312693 3.668474
#> 3                z 1.398680 0.9435062 0.1999072886 -1.0470808 3.844441 4.864887
#>   outcome
#> 1       y
#> 2       y
#> 3       y

# Subsetting works just as in `lm()`
lm_robust(y ~ x, data = dat, subset = z == 1)

#>   coefficient_name      est        se            p   ci_lower ci_upper df
#> 1      (Intercept) 4.941398 0.6493330 1.819289e-05  3.4945938 6.388202 10
#> 2                x 1.455851 0.7717047 8.856539e-02 -0.2636144 3.175316 10
#>   outcome
#> 1       y
#> 2       y

# One can also choose to set the significance level for different CIs
lm_robust(y ~ x + z, data = dat, alpha = 0.1)

#>   coefficient_name       est        se            p     ci_lower ci_upper df
#> 1      (Intercept) 3.8270190 0.4251936 7.460320e-11  3.109677710 4.544360 37
#> 2                x 0.5218375 0.3070137 9.757698e-02  0.003876662 1.039798 37
#> 3                z 1.0450491 0.7799145 1.884321e-01 -0.270739622 2.360838 37
#>   outcome
#> 1       y
#> 2       y
#> 3       y