Ordinary Least Squares with Robust Standard Errors

This formula fits a linear model, provides a variety of options for robust standard errors, and conducts coefficient tests

lm_robust(formula, data, weights, subset, clusters, se_type = NULL,
  ci = TRUE, alpha = 0.05, return_vcov = TRUE, try_cholesky = FALSE)

Arguments

formula
an object of class formula, as in lm

data
A data.frame

weights
the bare (unquoted) names of the weights variable in the supplied data.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

clusters
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.

se_type
The sort of standard error sought. If `clusters` is not specified the options are "HC0", "HC1" (or "stata", the equivalent), "HC2" (default), "HC3", or "classical". If `clusters` is specified the options are "CR0", "CR2" (default), or "stata". Can also specify "none", which may speed up estimation of the coefficients.

ci
logical. Whether to compute and return p-values and confidence intervals, TRUE by default.

alpha
The significance level, 0.05 by default.

return_vcov
logical. Whether to return the variance-covariance matrix for later usage, TRUE by default.

try_cholesky
logical. Whether to try using a Cholesky decomposition to solve least squares instead of a QR decomposition, FALSE by default. Using a Cholesky decomposition may result in speed gains, but should only be used if users are sure their model is full-rank (i.e., there is no perfect multi-collinearity)

Value

An object of class "lm_robust".

The post-estimation commands functions summary and tidy return results in a data.frame. To get useful data out of the return, you can use these data frames, you can use the resulting list directly, or you can use the generic accessor functions coef, vcov, confint, and predict. Marginal effects and uncertainty about them can be gotten by passing this object to margins from the margins.

Users who want to print the results in TeX of HTML can use the extract function and the texreg package.

If users specify a multivariate linear regression model (multiple outcomes), then some of the below components will be of higher dimension to accommodate the additional models.

An object of class "lm_robust" is a list containing at least the following components:

coefficients

the estimated coefficients

std.error

the estimated standard errors

df

the estimated degrees of freedom

p.value

the p-values from a two-sided t-test using coefficients, std.error, and df

ci.lower

the lower bound of the 1 - alpha percent confidence interval

ci.upper

the upper bound of the 1 - alpha percent confidence interval

term

a character vector of coefficient names

alpha

the significance level specified by the user

se_type

the standard error type specified by the user

res_var

the residual variance

N

the number of observations used

k

the number of columns in the design matrix (includes linearly dependent columns!)

rank

the rank of the fitted model

vcov

the fitted variance covariance matrix

r.squared

The \(R^2\), $$R^2 = 1 - Sum(e[i]^2) / Sum((y[i] - y^*)^2),$$ where \(y^*\) is the mean of \(y[i]\) if there is an intercept and zero otherwise, and \(e[i]\) is the ith residual.

adj.r.squared

The \(R^2\) but penalized for having more parameters, rank

fstatistic

a vector with the value of the F-statistic with the numerator and denominator degrees of freedom

weighted

whether or not weights were applied

call

the original function call

We also return terms and contrasts, used by predict.
Details

This function performs linear regression and provides a variety of standard errors. It takes a formula and data much in the same was as lm does, and all auxiliary variables, such as clusters and weights, can be passed either as quoted names of columns, as bare column names, or as a self-contained vector. Examples of usage can be seen below and in the Getting Started vignette.

The mathematical notes in this vignette specify the exact estimators used by this function. The default variance estimators have been chosen largely in accordance with the procedures in this manual. The default for the case without clusters is the HC2 estimator and the default with clusters is the analogous CR2 estimator. Users can easily replicate Stata standard errors in the clustered or non-clustered case by setting `se_type` = "stata".

The function estimates the coefficients and standard errors in C++, using the RcppEigen package. By default, we estimate the coefficients using Column-Pivoting QR decomposition from the Eigen C++ library, although users could get faster solutions by setting `try_cholesky` = TRUE to use a Cholesky decomposition instead. This will likely result in quicker solutions, but the algorithm does not reliably detect when there are linear dependencies in the model and may fail silently if they exist.

References

Abadie, Alberto, Susan Athey, Guido W Imbens, and Jeffrey Wooldridge. 2017. "A Class of Unbiased Estimators of the Average Treatment Effect in Randomized Experiments." arXiv Pre-Print. https://arxiv.org/abs/1710.02926v2.

Bell, Robert M, and Daniel F McCaffrey. 2002. "Bias Reduction in Standard Errors for Linear Regression with Multi-Stage Samples." Survey Methodology 28 (2): 169-82.

MacKinnon, James, and Halbert White. 1985. "Some Heteroskedasticity-Consistent Covariance Matrix Estimators with Improved Finite Sample Properties." Journal of Econometrics 29 (3): 305-25. https://doi.org/10.1016/0304-4076(85)90158-7.

Pustejovsky, James E, and Elizabeth Tipton. 2016. "Small Sample Methods for Cluster-Robust Variance Estimation and Hypothesis Testing in Fixed Effects Models." Journal of Business & Economic Statistics. Taylor & Francis. https://doi.org/10.1080/07350015.2016.1247004.

Samii, Cyrus, and Peter M Aronow. 2012. "On Equivalencies Between Design-Based and Regression-Based Variance Estimators for Randomized Experiments." Statistics and Probability Letters 82 (2). https://doi.org/10.1016/j.spl.2011.10.024.

Examples

library(fabricatr)
dat <- fabricate(
  N = 40,
  y = rpois(N, lambda = 4),
  x = rnorm(N),
  z = rbinom(N, 1, prob = 0.4)
)

# Default variance estimator is HC2 robust standard errors
lmro <- lm_robust(y ~ x + z, data = dat)

# Can tidy() the data in to a data.frame
tidy(lmro)

#>          term   estimate std.error      p.value   ci.lower  ci.upper df outcome
#> 1 (Intercept)  4.3437473 0.4266670 2.809933e-12  3.4792378 5.2082567 37       y
#> 2           x  0.1145175 0.2684168 6.721135e-01 -0.4293467 0.6583816 37       y
#> 3           z -0.7573433 0.6642250 2.615339e-01 -2.1031910 0.5885043 37       y

# Can use summary() to get more statistics
summary(lmro)

#> 
#> Call:
#> lm_robust(formula = y ~ x + z, data = dat)
#> 
#> Standard error type:  HC2 
#> 
#> Coefficients:
#>             Estimate Std. Error  Pr(>|t|) CI Lower CI Upper DF
#> (Intercept)   4.3437     0.4267 2.810e-12   3.4792   5.2083 37
#> x             0.1145     0.2684 6.721e-01  -0.4293   0.6584 37
#> z            -0.7573     0.6642 2.615e-01  -2.1032   0.5885 37
#> 
#> Multiple R-squared:  0.03794 ,	Adjusted R-squared:  -0.01407 
#> F-statistic: 0.7295 on 2 and 37 DF,  p-value: 0.4889

# Can also get coefficients three ways
lmro$coefficients

#> (Intercept)           x           z 
#>   4.3437473   0.1145175  -0.7573433 

coef(lmro)

#> (Intercept)           x           z 
#>   4.3437473   0.1145175  -0.7573433 

tidy(lmro)$estimate

#> [1]  4.3437473  0.1145175 -0.7573433

# Can also get confidence intervals from object or with new 1 - `alpha`
lmro$ci.lower

#> (Intercept)           x           z 
#>   3.4792378  -0.4293467  -2.1031910 

confint(lmro, level = 0.8)

#>                   10 %      90 %
#> (Intercept)  3.7870090 4.9004855
#> x           -0.2357274 0.4647623
#> z           -1.6240602 0.1093736

# Can recover classical standard errors
lmclassic <- lm_robust(y ~ x + z, data = dat, se_type = "classical")
tidy(lmclassic)

#>          term   estimate std.error      p.value   ci.lower  ci.upper df outcome
#> 1 (Intercept)  4.3437473 0.4252799 2.568950e-12  3.4820484 5.2054461 37       y
#> 2           x  0.1145175 0.2797434 6.846313e-01 -0.4522965 0.6813315 37       y
#> 3           z -0.7573433 0.6453147 2.480535e-01 -2.0648750 0.5501884 37       y

# Can easily match Stata's robust standard errors
lmstata <- lm_robust(y ~ x + z, data = dat, se_type = "stata")
tidy(lmstata)

#>          term   estimate std.error      p.value   ci.lower  ci.upper df outcome
#> 1 (Intercept)  4.3437473 0.4294666 3.362855e-12  3.4735653 5.2139292 37       y
#> 2           x  0.1145175 0.2638693 6.668137e-01 -0.4201325 0.6491675 37       y
#> 3           z -0.7573433 0.6628060 2.605316e-01 -2.1003159 0.5856293 37       y

# Easy to specify clusters for cluster-robust inference
dat$clusterID <- sample(1:10, size = 40, replace = TRUE)

lmclust <- lm_robust(y ~ x + z, data = dat, clusters = clusterID)
tidy(lmclust)

#>          term   estimate std.error      p.value  ci.lower  ci.upper       df
#> 1 (Intercept)  4.3437473 0.4464744 1.510243e-05  3.304129 5.3833652 7.578772
#> 2           x  0.1145175 0.3112151 7.282905e-01 -0.691010 0.9200449 4.888419
#> 3           z -0.7573433 0.6877050 3.103519e-01 -2.413794 0.8991070 6.419560
#>   outcome
#> 1       y
#> 2       y
#> 3       y

# Can also match Stata's clustered standard errors
lm_robust(
  y ~ x + z,
  data = dat,
  clusters = clusterID,
  se_type = "stata"
)

#>               Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper DF
#> (Intercept)  4.3437473  0.4405516 4.023991e-06  3.3471504 5.3403442  9
#> x            0.1145175  0.3031343 7.143509e-01 -0.5712199 0.8002548  9
#> z           -0.7573433  0.6575274 2.790786e-01 -2.2447737 0.7300871  9

# Works just as LM does with functions in the formula
dat$blockID <- rep(c("A", "B", "C", "D"), each = 10)

lm_robust(y ~ x + z + factor(blockID), data = dat)

#>                      Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper DF
#> (Intercept)       3.514890355  0.5772947 6.611384e-07  2.3416863 4.6880944 34
#> x                 0.006792508  0.2186290 9.753965e-01 -0.4375152 0.4511002 34
#> z                -0.778097499  0.6603591 2.468605e-01 -2.1201086 0.5639136 34
#> factor(blockID)B  1.320814446  0.9111379 1.563234e-01 -0.5308405 3.1724694 34
#> factor(blockID)C  1.774506204  0.7991346 3.315047e-02  0.1504694 3.3985430 34
#> factor(blockID)D  0.099756357  0.6652215 8.816826e-01 -1.2521365 1.4516492 34

# Weights are also easily specified
dat$w <- runif(40)

lm_robust(
  y ~ x + z,
  data = dat,
  weights = w,
  clusters = clusterID
)

#>                Estimate Std. Error     Pr(>|t|)  CI Lower CI Upper       DF
#> (Intercept)  4.02974602  0.4273344 2.518013e-05  3.026224 5.033268 7.247874
#> x           -0.08489078  0.3546075 8.258011e-01 -1.194862 1.025080 3.090917
#> z           -0.52984774  0.9463118 5.972174e-01 -2.886426 1.826730 5.596866

# Subsetting works just as in `lm()`
lm_robust(y ~ x, data = dat, subset = z == 1)

#>               Estimate Std. Error     Pr(>|t|)  CI Lower  CI Upper DF
#> (Intercept)  3.5392271  0.5544330 1.700539e-05  2.350087 4.7283677 14
#> x           -0.1114943  0.4294463 7.989333e-01 -1.032565 0.8095764 14

# One can also choose to set the significance level for different CIs
lm_robust(y ~ x + z, data = dat, alpha = 0.1)

#>               Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper DF
#> (Intercept)  4.3437473  0.4266670 2.809933e-12  3.6239201 5.0635744 37
#> x            0.1145175  0.2684168 6.721135e-01 -0.3383268 0.5673618 37
#> z           -0.7573433  0.6642250 2.615339e-01 -1.8779530 0.3632664 37

# NOT RUN {
  # Can also use 'margins' package if you have it installed to get
  # marginal effects
  library(margins)
  lmrout <- lm_robust(y ~ x + z, data = dat)
  summary(margins(lmrout))

  # Can output results using 'texreg'
  library(texreg)
  texreg(lmrout)
# }