Ordinary Least Squares with Robust Standard Errors

This formula fits a linear model, provides a variety of options for robust standard errors, and conducts coefficient tests

lm_robust(formula, data, weights, subset, clusters, se_type = NULL,
  ci = TRUE, alpha = 0.05, return_vcov = TRUE, try_cholesky = FALSE)

Arguments

formula
an object of class formula, as in lm

data
A data.frame

weights
the bare (unquoted) names of the weights variable in the supplied data.

subset
An optional bare (unquoted) expression specifying a subset of observations to be used.

clusters
An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.

se_type
The sort of standard error sought. If `clusters` is not specified the options are "HC0", "HC1" (or "stata", the equivalent), "HC2" (default), "HC3", or "classical". If `clusters` is specified the options are "CR0", "CR2" (default), or "stata". Can also specify "none", which may speed up estimation of the coefficients.

ci
logical. Whether to compute and return p-values and confidence intervals, TRUE by default.

alpha
The significance level, 0.05 by default.

return_vcov
logical. Whether to return the variance-covariance matrix for later usage, TRUE by default.

try_cholesky
logical. Whether to try using a Cholesky decomposition to solve least squares instead of a QR decomposition, FALSE by default. Using a Cholesky decomposition may result in speed gains, but should only be used if users are sure their model is full-rank (i.e., there is no perfect multi-collinearity)

Value

An object of class "lm_robust".

The post-estimation commands functions summary and tidy return results in a data.frame. To get useful data out of the return, you can use these data frames, you can use the resulting list directly, or you can use the generic accessor functions coef, vcov, confint, and predict. Marginal effects and uncertainty about them can be gotten by passing this object to margins from the margins.

Users who want to print the results in TeX of HTML can use the extract function and the texreg package.

If users specify a multivariate linear regression model (multiple outcomes), then some of the below components will be of higher dimension to accommodate the additional models.

An object of class "lm_robust" is a list containing at least the following components:

coefficients

the estimated coefficients

std.error

the estimated standard errors

df

the estimated degrees of freedom

p.value

the p-values from a two-sided t-test using coefficients, std.error, and df

ci.lower

the lower bound of the 1 - alpha percent confidence interval

ci.upper

the upper bound of the 1 - alpha percent confidence interval

term

a character vector of coefficient names

alpha

the significance level specified by the user

se_type

the standard error type specified by the user

res_var

the residual variance

N

the number of observations used

k

the number of columns in the design matrix (includes linearly dependent columns!)

rank

the rank of the fitted model

vcov

the fitted variance covariance matrix

r.squared

The \(R^2\), $$R^2 = 1 - Sum(e[i]^2) / Sum((y[i] - y^*)^2),$$ where \(y^*\) is the mean of \(y[i]\) if there is an intercept and zero otherwise, and \(e[i]\) is the ith residual.

adj.r.squared

The \(R^2\) but penalized for having more parameters, rank

fstatistic

a vector with the value of the F-statistic with the numerator and denominator degrees of freedom

weighted

whether or not weights were applied

call

the original function call

We also return terms and contrasts, used by predict.
Details

This function performs linear regression and provides a variety of standard errors. It takes a formula and data much in the same was as lm does, and all auxiliary variables, such as clusters and weights, can be passed either as quoted names of columns, as bare column names, or as a self-contained vector. Examples of usage can be seen below and in the Getting Started vignette.

The mathematical notes in this vignette specify the exact estimators used by this function. The default variance estimators have been chosen largely in accordance with the procedures in this manual. The default for the case without clusters is the HC2 estimator and the default with clusters is the analogous CR2 estimator. Users can easily replicate Stata standard errors in the clustered or non-clustered case by setting `se_type` = "stata".

The function estimates the coefficients and standard errors in C++, using the RcppEigen package. By default, we estimate the coefficients using Column-Pivoting QR decomposition from the Eigen C++ library, although users could get faster solutions by setting `try_cholesky` = TRUE to use a Cholesky decomposition instead. This will likely result in quicker solutions, but the algorithm does not reliably detect when there are linear dependencies in the model and may fail silently if they exist.

References

Abadie, Alberto, Susan Athey, Guido W Imbens, and Jeffrey Wooldridge. 2017. "A Class of Unbiased Estimators of the Average Treatment Effect in Randomized Experiments." arXiv Pre-Print. https://arxiv.org/abs/1710.02926v2.

Bell, Robert M, and Daniel F McCaffrey. 2002. "Bias Reduction in Standard Errors for Linear Regression with Multi-Stage Samples." Survey Methodology 28 (2): 169-82.

MacKinnon, James, and Halbert White. 1985. "Some Heteroskedasticity-Consistent Covariance Matrix Estimators with Improved Finite Sample Properties." Journal of Econometrics 29 (3): 305-25. https://doi.org/10.1016/0304-4076(85)90158-7.

Pustejovsky, James E, and Elizabeth Tipton. 2016. "Small Sample Methods for Cluster-Robust Variance Estimation and Hypothesis Testing in Fixed Effects Models." Journal of Business & Economic Statistics. Taylor & Francis. https://doi.org/10.1080/07350015.2016.1247004.

Samii, Cyrus, and Peter M Aronow. 2012. "On Equivalencies Between Design-Based and Regression-Based Variance Estimators for Randomized Experiments." Statistics and Probability Letters 82 (2). https://doi.org/10.1016/j.spl.2011.10.024.

Examples

library(fabricatr)
dat <- fabricate(
  N = 40,
  y = rpois(N, lambda = 4),
  x = rnorm(N),
  z = rbinom(N, 1, prob = 0.4)
)

# Default variance estimator is HC2 robust standard errors
lmro <- lm_robust(y ~ x + z, data = dat)

# Can tidy() the data in to a data.frame
tidy(lmro)

#>          term     estimate std.error      p.value   ci.lower  ci.upper df
#> 1 (Intercept)  4.199089608 0.3806467 2.955510e-13  3.4278260 4.9703532 37
#> 2           x -0.003271224 0.2654557 9.902341e-01 -0.5411357 0.5345932 37
#> 3           z -0.133687393 0.7152674 8.527559e-01 -1.5829569 1.3155821 37
#>   outcome
#> 1       y
#> 2       y
#> 3       y

# Can use summary() to get more statistics
summary(lmro)

#> 
#> Call:
#> lm_robust(formula = y ~ x + z, data = dat)
#> 
#> Standard error type:  HC2 
#> 
#> Coefficients:
#>              Estimate Std. Error  Pr(>|t|) CI Lower CI Upper DF
#> (Intercept)  4.199090     0.3806 2.956e-13   3.4278   4.9704 37
#> x           -0.003271     0.2655 9.902e-01  -0.5411   0.5346 37
#> z           -0.133687     0.7153 8.528e-01  -1.5830   1.3156 37
#> 
#> Multiple R-squared:  0.001038 ,	Adjusted R-squared:  -0.05296 
#> F-statistic: 0.01922 on 2 and 37 DF,  p-value: 0.981

# Can also get coefficients three ways
lmro$coefficients

#>  (Intercept)            x            z 
#>  4.199089608 -0.003271224 -0.133687393 

coef(lmro)

#>  (Intercept)            x            z 
#>  4.199089608 -0.003271224 -0.133687393 

tidy(lmro)$estimate

#> [1]  4.199089608 -0.003271224 -0.133687393

# Can also get confidence intervals from object or with new 1 - `alpha`
lmro$ci.lower

#> (Intercept)           x           z 
#>   3.4278260  -0.5411357  -1.5829569 

confint(lmro, level = 0.8)

#>                   10 %      90 %
#> (Intercept)  3.7024010 4.6957782
#> x           -0.3496523 0.3431099
#> z           -1.0670072 0.7996324

# Can recover classical standard errors
lmclassic <- lm_robust(y ~ x + z, data = dat, se_type = "classical")
tidy(lmclassic)

#>          term     estimate std.error      p.value   ci.lower  ci.upper df
#> 1 (Intercept)  4.199089608 0.4249397 6.341373e-12  3.3380801 5.0600992 37
#> 2           x -0.003271224 0.2917437 9.911140e-01 -0.5944001 0.5878576 37
#> 3           z -0.133687393 0.6818706 8.456362e-01 -1.5152884 1.2479136 37
#>   outcome
#> 1       y
#> 2       y
#> 3       y

# Can easily match Stata's robust standard errors
lmstata <- lm_robust(y ~ x + z, data = dat, se_type = "stata")
tidy(lmstata)

#>          term     estimate std.error      p.value   ci.lower  ci.upper df
#> 1 (Intercept)  4.199089608 0.3847691 4.018442e-13  3.4194734 4.9787058 37
#> 2           x -0.003271224 0.2628282 9.901365e-01 -0.5358117 0.5292693 37
#> 3           z -0.133687393 0.7121457 8.521181e-01 -1.5766316 1.3092568 37
#>   outcome
#> 1       y
#> 2       y
#> 3       y

# Easy to specify clusters for cluster-robust inference
dat$clusterID <- sample(1:10, size = 40, replace = TRUE)

lmclust <- lm_robust(y ~ x + z, data = dat, clusters = clusterID)
tidy(lmclust)

#>          term     estimate std.error      p.value   ci.lower  ci.upper       df
#> 1 (Intercept)  4.199089608 0.2830495 1.169007e-06  3.5334908 4.8646884 7.197601
#> 2           x -0.003271224 0.2836576 9.911693e-01 -0.6958089 0.6892665 6.055808
#> 3           z -0.133687393 0.7231267 8.586410e-01 -1.8484317 1.5810569 6.904203
#>   outcome
#> 1       y
#> 2       y
#> 3       y

# Can also match Stata's clustered standard errors
lm_robust(
  y ~ x + z,
  data = dat,
  clusters = clusterID,
  se_type = "stata"
)

#>                 Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper DF
#> (Intercept)  4.199089608  0.2823965 1.217032e-07  3.5602644 4.8379148  9
#> x           -0.003271224  0.2776595 9.908570e-01 -0.6313807 0.6248383  9
#> z           -0.133687393  0.7051469 8.538388e-01 -1.7288404 1.4614656  9

# Works just as LM does with functions in the formula
dat$blockID <- rep(c("A", "B", "C", "D"), each = 10)

lm_robust(y ~ x + z + factor(blockID), data = dat)

#>                     Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper DF
#> (Intercept)       3.60051619  0.4748595 8.293432e-09  2.6354856 4.5655468 34
#> x                -0.01055316  0.2377445 9.648543e-01 -0.4937082 0.4726019 34
#> z                -0.01244249  0.6899951 9.857182e-01 -1.4146812 1.3897963 34
#> factor(blockID)B  1.19721445  0.8362142 1.613566e-01 -0.5021772 2.8966061 34
#> factor(blockID)C  1.50355320  0.7924818 6.631185e-02 -0.1069636 3.1140700 34
#> factor(blockID)D -0.49762991  0.7061475 4.857898e-01 -1.9326944 0.9374345 34

# Weights are also easily specified
dat$w <- runif(40)

lm_robust(
  y ~ x + z,
  data = dat,
  weights = w,
  clusters = clusterID
)

#>                Estimate Std. Error     Pr(>|t|)  CI Lower  CI Upper       DF
#> (Intercept)  3.90335473  0.2997135 5.892220e-06  3.186038 4.6206712 6.608244
#> x           -0.26916114  0.2959378 4.043239e-01 -1.027231 0.4889091 5.059028
#> z            0.01175879  0.8200111 9.890292e-01 -2.000542 2.0240592 5.929398

# Subsetting works just as in `lm()`
lm_robust(y ~ x, data = dat, subset = z == 1)

#>              Estimate Std. Error     Pr(>|t|)   CI Lower CI Upper DF
#> (Intercept) 4.2100287  0.6040706 9.777948e-06  2.9050135 5.515044 13
#> x           0.3708874  0.4544886 4.291690e-01 -0.6109757 1.352750 13

# One can also choose to set the significance level for different CIs
lm_robust(y ~ x + z, data = dat, alpha = 0.1)

#>                 Estimate Std. Error     Pr(>|t|)   CI Lower  CI Upper DF
#> (Intercept)  4.199089608  0.3806467 2.955510e-13  3.5569029 4.8412763 37
#> x           -0.003271224  0.2654557 9.902341e-01 -0.4511199 0.4445775 37
#> z           -0.133687393  0.7152674 8.527559e-01 -1.3404105 1.0730357 37

# NOT RUN {
  # Can also use 'margins' package if you have it installed to get
  # marginal effects
  library(margins)
  lmrout <- lm_robust(y ~ x + z, data = dat)
  summary(margins(lmrout))

  # Can output results using 'texreg'
  library(texreg)
  texregobj <- extract(lmrout)
# }