Internal method that creates linear fits

Internal method that creates linear fits

lm_robust_fit(y, X, weights, cluster, ci, se_type, alpha, coefficient_name,
  return_vcov)

Arguments

y
numeric outcome vector

X
numeric design matrix

weights
numeric weights vector

cluster
numeric cluster vector

ci
boolean that when T returns confidence intervals and p-values

se_type
character denoting which kind of SEs to return

alpha
numeric denoting the test size for confidence intervals

coefficient_name
character vector of coefficients to return

return_vcov
a boolean for whether to return the vcov matrix for later usage